refactor(services): extract more pure helpers from large services
- StorageAccessService.kt: 1781 -> 1186 LOC. New sibling files:
  * StorageMetadataDerivers.kt (310 LOC) — file-name and EPUB/PDF/comic-archive metadata derivation, OPF parsing, PDF info extraction
  * StorageDocumentFileHelpers.kt (128 LOC) — DocumentFile copy/move helpers + unique-name picker
  * StorageImportLogHelpers.kt (44 LOC) — read/write/list import transaction logs
  * StorageImportPlanBuilder.kt (161 LOC) — buildPlanRecursively + computeSha256 + deleteEmptyDirectories
  Remaining oversize is the orchestrator import methods (executeImportPlanAdvanced, importRecursively) which depend heavily on injected sanitizer/queues and need real architectural surgery.

- FanfictionEpubConversionService.kt: 933 -> 501 LOC. New sibling:
  * FanfictionEpubBuilder.kt (505 LOC) — EPUB packaging (container.xml, content.opf, TOC, title page, chapter XHTML, stylesheet), HTML cleaning, XML escaping, CRC32, FicHub/FFNet/AO3/Wattpad/RoyalRoad chapter & metadata parsers.

- InternetRadioViewModel.kt: 901 -> 127 LOC. New sibling:
  * SampleRadioStations.kt (780 LOC) — the 777-LOC hard-coded station catalog moves to buildSampleStations(). The ViewModel now calls the helper to populate _sampleStations.
