+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/radio/InternetRadioViewModel.kt
@@ -75,781 +75,7 @@ class InternetRadioViewModel @Inject constructor(
     }
 
     private fun loadSampleStations() {
-        // Comprehensive collection from 4 premium sources
-        // Total: 504 unique high-quality radio stations!
-        // (751 declarations, 504 unique URLs after deduplication)
-        _sampleStations.value = listOf(
-            // ==================================================================
-            // filtermusic.net Collection (241 stations)
-            // Source: https://filtermusic.net/
-            // ==================================================================
-InternetRadioStation("fm1", "Deepershades", "https://andromeda.housejunkie.ca/radio/8000/radio.mp3", "House", "128 kbps"),
-InternetRadioStation("fm2", "Sound of Berlin", "https://fluxmusic.api.radiosphere.io/channels/sound-of-berlin/stream.aac", "House", "128 kbps"),
-InternetRadioStation("fm3", "Ibiza Global", "https://cdn-peer031.streaming-pro.com:8025/ibizaglobalradio.mp3", "House", "128 kbps"),
-InternetRadioStation("fm4", "Dogglounge", "https://securestreams4.autopo.st:1853/stream", "House", "128 kbps"),
-InternetRadioStation("fm5", "Vanilla Radio", "https://stream.vanillaradio.com:8003/stream", "House", "128 kbps"),
-InternetRadioStation("fm6", "Cavo Paradiso", "https://neos.win:48488/1", "House", "128 kbps"),
-InternetRadioStation("fm7", "Berlin Beach House Radio", "https://fluxmusic.api.radiosphere.io/channels/berlin-beach-house/stream.aac", "House", "128 kbps"),
-InternetRadioStation("fm8", "Deep House Radio", "https://streaming.shoutcast.com/dhr?1711824191948", "House", "128 kbps"),
-InternetRadioStation("fm9", "Club Sandwich", "https://streams.fluxfm.de/clubsandwich/mp3-128/streams.fluxfm.de", "House", "128 kbps"),
-InternetRadioStation("fm10", "Point Blank", "https://pointblankradio.co.uk/stream", "House", "128 kbps"),
-InternetRadioStation("fm11", "Deepinside", "https://stream.rcs.revma.com/uyrbt6xuhnruv", "House", "128 kbps"),
-InternetRadioStation("fm12", "DinamoFM Deep", "https://channels.dinamo.fm/deep-mp3", "House", "128 kbps"),
-InternetRadioStation("fm13", "Radio Deep", "https://s2.radio.co/s098f4886e/listen", "House", "128 kbps"),
-InternetRadioStation("fm14", "Ibiza Sonica Club", "https://ibizasonica.streaming-pro.com:8011/sonicaclub", "House", "128 kbps"),
-InternetRadioStation("fm15", "Mixadance", "https://stream.mixadance.fm/mixadance", "House", "128 kbps"),
-InternetRadioStation("fm16", "FG Avant Garde", "https://n11.rcs.revma.com/02s8zcv9898uv", "House", "128 kbps"),
-InternetRadioStation("fm17", "Ibiza Live Radio", "https://uksoutha.streaming.broadcast.radio/ibiza-live-radio", "House", "128 kbps"),
-InternetRadioStation("fm18", "Barcelona City fm", "https://s2.radio.co/sd9da54aa0/listen", "House", "128 kbps"),
-InternetRadioStation("fm19", "Physical", "https://s3.radio.co/sd59eccabf/listen", "House", "128 kbps"),
-InternetRadioStation("fm20", "Deep Radio", "https://29043.live.streamtheworld.com/DEEP_RADIOAAC.aac", "House", "128 kbps"),
-InternetRadioStation("fm21", "Experimental TV Radio", "https://s44.myradiostream.com/8890/listen.mp3", "House", "128 kbps"),
-InternetRadioStation("fm22", "Housefm", "https://uksoutha.streaming.broadcast.radio/housefm", "House", "128 kbps"),
-InternetRadioStation("fm23", "DinamoFM", "https://channels.dinamo.fm/legacy-mp3", "House", "128 kbps"),
-InternetRadioStation("fm24", "Frisky Deep", "https://stream.deep.friskyradio.com/mp3_low", "House", "128 kbps"),
-InternetRadioStation("fm25", "Music Factory", "https://i4.streams.ovh:2200/ssl/musicfactory?mp=/stream", "House", "128 kbps"),
-InternetRadioStation("fm26", "Fantasy FM", "https://eu4.fastcast4u.com/proxy/fantasyfm?mp=/1", "House", "128 kbps"),
-InternetRadioStation("fm27", "Deepinradio", "https://ssl1.viastreaming.net:8525/stream.mp3", "House", "128 kbps"),
-InternetRadioStation("fm28", "MoveDaHouse", "https://uk7.internet-radio.com/proxy/movedahouse?mp=//stream", "House", "128 kbps"),
-InternetRadioStation("fm29", "We get lifted", "https://s3.radio.co/sa80d22794/listen", "House", "128 kbps"),
-InternetRadioStation("fm30", "Mambo", "https://streamer.radio.co/sd7b28e5f3/listen", "House", "128 kbps"),
-InternetRadioStation("fm31", "Blue Marlin", "https://ibizasonica.streaming-pro.com:8001/bluemarlin", "House", "128 kbps"),
-InternetRadioStation("fm32", "MotionFM House", "https://vm.motionfm.com/motionone_free", "House", "128 kbps"),
-InternetRadioStation("fm33", "320fm", "https://stream.weblygo.de:8020/stream", "House", "128 kbps"),
-InternetRadioStation("fm34", "Centreforce", "https://listen.centreforceradio.com:8830/192", "House", "128 kbps"),
-InternetRadioStation("fm35", "BeachGrooves", "https://stream.beachgrooves.com:9000/;*.mp3", "House", "128 kbps"),
-InternetRadioStation("fm36", "Freak31", "https://stream.31media.net/freak31", "House", "128 kbps"),
-InternetRadioStation("fm37", "1 Radio Space ", "https://c22.radioboss.fm:18118/1RADIO.SPACE", "Techno", "128 kbps"),
-InternetRadioStation("fm38", "Frisky", "https://stream.frisky.friskyradio.com/mp3_low", "Techno", "128 kbps"),
-InternetRadioStation("fm39", "Insomnia fm", "https://insomniafm.cloudrad.io/live", "Techno", "128 kbps"),
-InternetRadioStation("fm40", "Proton radio", "https://shoutcast.protonradio.com/;*.mp3", "Techno", "128 kbps"),
-InternetRadioStation("fm41", "Joint Beat", "https://jointil.com/stream-beat", "Techno", "128 kbps"),
-InternetRadioStation("fm42", "TrancePulse", "https://stream.trance.ie:8443/stream", "Techno", "128 kbps"),
-InternetRadioStation("fm43", "Afterhours fm", "https://fr.ah.fm/live", "Techno", "128 kbps"),
-InternetRadioStation("fm44", "Trance Athena", "https://cast.streams.ovh:8008/;*.mp3", "Techno", "128 kbps"),
-InternetRadioStation("fm45", "Psyland", "https://sp.streams.ovh/8020/stream", "Techno", "128 kbps"),
-InternetRadioStation("fm46", "Discover trance", "https://uk01.discovertrance.com/;*.mp3", "Techno", "128 kbps"),
-InternetRadioStation("fm47", "FIP electro", "https://icecast.radiofrance.fr/fipelectro-midfi.mp3", "Electronic", "128 kbps"),
-InternetRadioStation("fm48", "Intergalactic", "https://radio.intergalactic.fm/1", "Electronic", "128 kbps"),
-InternetRadioStation("fm49", "Rinse UK", "https://admin.stream.rinse.fm/proxy/rinse_uk/stream", "Electronic", "128 kbps"),
-InternetRadioStation("fm50", "Sovietwave", "https://station.waveradio.org/soviet", "Electronic", "128 kbps"),
-InternetRadioStation("fm51", "Plusfm", "https://stream.plusfm.net/1", "Electronic", "128 kbps"),
-InternetRadioStation("fm52", "Biptunia", "https://ecast.myautodj.com:1380/listen.mp3", "Electronic", "128 kbps"),
-InternetRadioStation("fm53", "Noise FM", "https://play.sas-media.ru/play_256", "Electronic", "128 kbps"),
-InternetRadioStation("fm54", "Systrum Sistum Electronica", "https://systrum.net:8443/SSR2", "Electronic", "128 kbps"),
-InternetRadioStation("fm55", "Dirty", "https://live.dirty.radio/stream", "Electronic", "128 kbps"),
-InternetRadioStation("fm56", "AmpedOut", "https://s12.ssl-stream.com/8006/stream", "Electronic", "128 kbps"),
-InternetRadioStation("fm57", "Bloodlit", "https://usa5.fastcast4u.com/proxy/wwwblood?mp=/1", "Electronic", "128 kbps"),
-InternetRadioStation("fm58", "Bassdrive", "https://bassdrive.radioca.st/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm59", "Renegade", "https://streaming06.liveboxstream.uk/proxy/renegade/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm60", "Jungletrain", "https://chat.jungletrain.net/streamtest/;stream/1", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm61", "SubFM", "https://subfm.radioca.st/live", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm62", "Breakbeat Zone", "https://radio.bbz.ru/listen/bbz/192.mp3", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm63", "Dirty Mayhem dnb", "https://node-34.zeno.fm/x7d910w23d0uv", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm64", "Sunshine Live Drum-n-Bass", "https://sunsl.streamabc.net/sunsl-drumnbass-mp3-192-9834247", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm65", "UKBassRadio", "https://s2.ssl-stream.com/listen/uk_bass_radio/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm66", "Dnbradio", "https://dnbradio.nl/dnbradio_main.mp3", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm67", "Old Skool Anthemz", "https://osa.radioca.st/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm68", "Laylow FM", "https://stream.laylowlondon.co.uk/listen/laylowfm/radio.ogg", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm69", "Dirtybass", "https://kentha.dirtybass.fm/endpoint/dirtybass.mp3", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm70", "Partyvibe Drum and Bass", "https://www.partyviberadio.com:8061", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm71", "WeFunk", "https://s-09.wefunkradio.com:8443/wefunk64.mp3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm72", "Badradio", "https://s2.radio.co/s2b2b68744/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm73", "FIP hiphop", "https://icecast.radiofrance.fr/fiphiphop-midfi.mp3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm74", "HipHop Classics", "https://fluxmusic.api.radiosphere.io/channels/boom-fm-classics/stream.aac", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm75", "Le Rapologue", "https://le-rapologue-radio.radiocult.fm/stream", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm76", "Tequila HipHop", "https://ssl.omegahost.ro/7000/stream", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm77", "DTLR", "https://radio.shoutcheap.com/proxy/shcaesar/stream.mp3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm78", "Itch fm", "https://streaming.radio.co/s264858a04/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm79", "Dilemaradio", "https://dilemaradiolive.radioca.st/stream", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm80", "Base FM", "https://stream.basefm.co.nz/Base320MP3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm81", "Hot 108 Jamz", "https://live.powerhitz.com/hot108", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm82", "Scratchvision", "https://stream.radio.co/saacde1bcb/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm83", "El Bloke", "https://streaming.radio.co/sd8f1e3a16/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm84", "Joint", "https://jointil.com/stream-reggae", "Reggae", "128 kbps"),
-InternetRadioStation("fm85", "Blue radio", "https://a3.asurahosting.com:7960/radio.mp3", "Reggae", "128 kbps"),
-InternetRadioStation("fm86", "TikiPod", "https://cast1.torontocast.com:3330/stream", "Reggae", "128 kbps"),
-InternetRadioStation("fm87", "Roots Legacy", "https://l.rootslegacy.fr/stream", "Reggae", "128 kbps"),
-InternetRadioStation("fm88", "Pigpen", "https://azuracast.spollin.net:8000/radio.mp3", "Reggae", "128 kbps"),
-InternetRadioStation("fm89", "Partyvibe reggae", "https://www.partyviberadio.com:8060", "Reggae", "128 kbps"),
-InternetRadioStation("fm90", "TBN Reggae", "https://puma.streemlion.com:1085/stream", "Reggae", "128 kbps"),
-InternetRadioStation("fm91", "Jah Music Mansion", "https://usa13.fastcast4u.com/proxy/jahmusicmansion?mp=/1", "Reggae", "128 kbps"),
-InternetRadioStation("fm92", "B4B Funk", "https://eu10.fastcast4u.com:8120/;*.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm93", "B-Funk", "https://streams.fluxfm.de/event01/mp3-128/streams.fluxfm.de", "Funk", "128 kbps"),
-InternetRadioStation("fm94", "Soulconnexion", "https://live.soulconnexionradio.com:1045/stream", "Funk", "128 kbps"),
-InternetRadioStation("fm95", "Funky Corner", "https://ais-sa2.cdnstream1.com/2447_192.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm96", "Soul Legends", "https://streamer.radio.co/s60f3e59d1/listen", "Funk", "128 kbps"),
-InternetRadioStation("fm97", "Funky", "https://funkyradio.streamingmedia.it/play.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm98", "Disco Paradise", "https://broadcast.miami/proxy/thediscoparadise?mp=/stream/;", "Funk", "128 kbps"),
-InternetRadioStation("fm99", "Nula", "https://strm.radionula.com/classics", "Funk", "128 kbps"),
-InternetRadioStation("fm100", "Goodfellas Music Station", "https://onair7.xdevel.com/proxy/xautocloud_q1v0_1005?mp=/;stream/", "Funk", "128 kbps"),
-InternetRadioStation("fm101", "Disco fetish", "https://radio.intergalactic.fm/2", "Funk", "128 kbps"),
-InternetRadioStation("fm102", "Luxfunk Blackmix", "https://luxfunkbroadcast.com/proxy/luxfunkradio/stream", "Funk", "128 kbps"),
-InternetRadioStation("fm103", "Palace", "https://www.radioking.com/play/palace-radio", "Funk", "128 kbps"),
-InternetRadioStation("fm104", "Generation Soul Disco Funk", "https://gestream.fr/g-radio-hd.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm105", "DanceGroove", "https://server7.streamserver24.com:8080/proxy/marasalf?mp=/stream", "Funk", "128 kbps"),
-InternetRadioStation("fm106", "Nicecream Green", "https://play.nicecream.fm/radio/8010/green.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm107", "Bless your sound", "https://ibizasonica.streaming-pro.com:8015/nomorenoless", "Lounge", "128 kbps"),
-InternetRadioStation("fm108", "Cafe Del Mar", "https://streams.radio.co/se1a320b47/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm109", "Croissant Show", "https://streams.radio.co/sddfa96b70/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm110", "Nicecream Red", "https://play.nicecream.fm/radio/8000/red.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm111", "Jazz de Ville Groove", "https://onair22.xdevel.com/proxy/xautocloud_1kha_423?mp=/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm112", "Qmusica", "https://eu8.fastcast4u.com/proxy/qfmserver2?mp=/1", "Lounge", "128 kbps"),
-InternetRadioStation("fm113", "FluxLounge", "https://streams.fluxfm.de/lounge/mp3-128/streams.fluxfm.de", "Lounge", "128 kbps"),
-InternetRadioStation("fm114", "RadioActive", "https://streamyourdream.org:8050/radioactive", "Lounge", "128 kbps"),
-InternetRadioStation("fm115", "Jockey", "https://ec4.yesstreaming.net:2135/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm116", "MotionFM Lounge", "https://vm.motionfm.com/motionthree_aacp", "Lounge", "128 kbps"),
-InternetRadioStation("fm117", "Vanilla Smooth", "https://smooth.vanillaradio.com:8032/live", "Lounge", "128 kbps"),
-InternetRadioStation("fm118", "Nordic Lodge", "https://radio.streemlion.com:3590/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm119", "Crooner", "https://croonerradio.ice.infomaniak.ch/croonerradio-midfi.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm120", "Lounge Radio", "https://nl1.streamhosting.ch/lounge128.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm121", "Salalondon", "https://solid55.streamupsolutions.com/proxy/cfyzzdah/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm122", "Sonicalm", "https://ibizasonica.streaming-pro.com:8014/sonicalm", "Lounge", "128 kbps"),
-InternetRadioStation("fm123", "Skylab", "https://uksoutha.streaming.broadcast.radio:29690/skylab-radio-limited", "Lounge", "128 kbps"),
-InternetRadioStation("fm124", "DinamoFM Caffe", "https://channels.dinamo.fm/caffe-mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm125", "Skyline radio e soul", "https://rblive.it:8010/radio.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm126", "ABC Lounge", "https://listen.openstream.co/3139/audio", "Lounge", "128 kbps"),
-InternetRadioStation("fm127", "Luxuriamusic", "https://ice10.securenetsystems.net/LUXOMP3", "Lounge", "128 kbps"),
-InternetRadioStation("fm128", "Melodies", "https://i.streams.ovh:2200/ssl/melodies?mp=/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm129", "Bellostream", "https://stream.radio.co/s0ac307198/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm130", "OneLuvFM", "https://streamer.radio.co/s34d8d1549/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm131", "Smoothlounge", "https://smoothjazz.cdnstream1.com/2586_128.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm132", "Maretimo Lounge", "https://s35.derstream.net/lounge.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm133", "SwissGroove", "https://relay1.swissgroove.ch/;*.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm134", "Systrum Sistum", "https://systrum.net:8443/SSR1", "Ambient", "128 kbps"),
-InternetRadioStation("fm135", "Ambient Radio", "https://uk2.internet-radio.com/proxy/ambientradio?mp=/;", "Ambient", "128 kbps"),
-InternetRadioStation("fm136", "ChillHop by FluxFM", "https://streams.fluxfm.de/Chillhop/mp3-128/streams.fluxfm.de", "Ambient", "128 kbps"),
-InternetRadioStation("fm137", "Yoga Sounds by FluxFM", "https://streams.fluxfm.de/yogasounds/mp3-128/streams.fluxfm.de", "Ambient", "128 kbps"),
-InternetRadioStation("fm138", "Verdure", "https://verdure.net/radio/192", "Ambient", "128 kbps"),
-InternetRadioStation("fm139", "Journeyscapes", "https://eagle.streemlion.com/proxy/journeys/stream", "Ambient", "128 kbps"),
-InternetRadioStation("fm140", "Psyndora", "https://cast.magicstreams.gr/sc/psychill/stream", "Ambient", "128 kbps"),
-InternetRadioStation("fm141", "Meuh", "https://radiomeuh2.ice.infomaniak.ch/radiomeuh2-128.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm142", "Noods", "https://noods-radio.radiocult.fm/stream", "Music", "128 kbps"),
-InternetRadioStation("fm143", "Kiosk", "https://kioskradiobxl.out.airtime.pro/kioskradiobxl_b", "Music", "128 kbps"),
-InternetRadioStation("fm144", "Stegi Radio - The Other Side", "https://movementathens2.out.airtime.pro/movementathens2_a", "Music", "128 kbps"),
-InternetRadioStation("fm145", "Le Son Parisien", "https://stream.lesonparisien.com/hi.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm146", "Athens Voice Electronic", "https://n09.radiojar.com/u52ufw1qayduv", "Music", "128 kbps"),
-InternetRadioStation("fm147", "Open Broadcast", "https://stream.openbroadcast.ch/320.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm148", "The Music Machine", "https://msmn3.co/proxy/x2mp33/stream", "Music", "128 kbps"),
-InternetRadioStation("fm149", "OpenLab", "https://ice04.fluidstream.net/openlab.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm150", "Monkey Bros", "https://n0b.radiojar.com/5tyqrwpgpfeuv", "Music", "128 kbps"),
-InternetRadioStation("fm151", "Raheem", "https://radioraheem.out.airtime.pro/radioraheem_a", "Music", "128 kbps"),
-InternetRadioStation("fm152", "Space dogs", "https://www.radioking.com/play/space-dogs-radio", "Music", "128 kbps"),
-InternetRadioStation("fm153", "Vintage Obscura", "https://radio.vintageobscura.net/stream", "Music", "128 kbps"),
-InternetRadioStation("fm154", "DinamoFM Smog", "https://channels.dinamo.fm/smog-mp3", "Music", "128 kbps"),
-InternetRadioStation("fm155", "Stegi Radio", "https://movementathens.out.airtime.pro/movementathens_a", "Music", "128 kbps"),
-InternetRadioStation("fm156", "Music Box", "https://uk5.internet-radio.com/proxy/musicboxradio?mp=/stream", "Music", "128 kbps"),
-InternetRadioStation("fm157", "Global Funk Radio", "https://securestreams2.autopo.st:1148/;*.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm158", "Movie Radio", "https://stream.klassikradio.de/movie-national/aac-128/konsoleplayer", "Music", "128 kbps"),
-InternetRadioStation("fm159", "I Have A Dream", "https://live.radio-ihaveadream.com/stream", "Music", "128 kbps"),
-InternetRadioStation("fm160", "Radio 3S", "https://radio.solarsoundsystem.org/listen/radio3s/radio.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm161", "Worldwide FM", "https://worldwidefm.out.airtime.pro/worldwidefm_b", "Music", "128 kbps"),
-InternetRadioStation("fm162", "Kiss FM", "https://s9.nexuscast.com:9565/kissfm.aac", "Music", "128 kbps"),
-InternetRadioStation("fm163", "Futuradios", "https://futuradiohits.ice.infomaniak.ch/frhits-128.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm164", "Swisspop", "https://stream.srg-ssr.ch/m/rsp/mp3_128", "Music", "128 kbps"),
-InternetRadioStation("fm165", "Maxxima", "https://maxxima.mine.nu/maxxima.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm166", "Frequence3", "https://frequence3.net-radio.fr/frequence3-128.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm167", "PurelectroRadio", "https://str0.creacast.com/purelectro", "Music", "128 kbps"),
-InternetRadioStation("fm168", "EDM Sessions", "https://s2.radio.co/s30844a0f4/listen", "Music", "128 kbps"),
-InternetRadioStation("fm169", "GBH Music", "https://rocket.streamradio.fr:8062/stream", "Music", "128 kbps"),
-InternetRadioStation("fm170", "Tsugi", "https://www.radioking.com/play/tsugi-radio", "Music", "128 kbps"),
-InternetRadioStation("fm171", "8radio", "https://securestreams.autopo.st:1035/;*.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm172", "Le Village Pop", "https://www.radioking.com/play/levillagepop", "Music", "128 kbps"),
-InternetRadioStation("fm173", "Beatles radio", "https://strw3.openstream.co/981", "Music", "128 kbps"),
-InternetRadioStation("fm174", "Dream Machine", "https://radio.intergalactic.fm/3", "Music", "128 kbps"),
-InternetRadioStation("fm175", "NTS", "https://stream-relay-geo.ntslive.net/stream", "Music", "128 kbps"),
-InternetRadioStation("fm176", "17bis", "https://listen.radioking.com/radio/444/stream/1754", "Music", "128 kbps"),
-InternetRadioStation("fm177", "EFR12", "https://listen.radioking.com/radio/151768/stream/192217", "Music", "128 kbps"),
-InternetRadioStation("fm178", "Dublab", "https://dublab.out.airtime.pro/dublab_a", "Music", "128 kbps"),
-InternetRadioStation("fm179", "GlitterBeam", "https://ukwesta.streaming.broadcast.radio/glitterbeam", "Music", "128 kbps"),
-InternetRadioStation("fm180", "Explorers Emporium", "https://www.radioking.com/play/explorers-emporium-radio", "Music", "128 kbps"),
-InternetRadioStation("fm181", "Electro Swing Revolution", "https://streamer.radio.co/s2c3cc784b/listen", "Music", "128 kbps"),
-InternetRadioStation("fm182", "Planet Pootwaddle", "https://ppw.streamguys1.com/sgplayer-mp3", "Music", "128 kbps"),
-InternetRadioStation("fm183", "80s Forever", "https://premium.shoutcastsolutions.com/radio/8050/256.mp3", "Retro", "128 kbps"),
-InternetRadioStation("fm184", "Lolli Oldies", "https://stream.lolliradio.net/lolli_oldies.mp3", "Retro", "128 kbps"),
-InternetRadioStation("fm185", "Fantasy radio", "https://italo.italo.nu/live", "Retro", "128 kbps"),
-InternetRadioStation("fm186", "Groovy Reflections", "https://usa12.fastcast4u.com/proxy/reflecti?mp=/1", "Retro", "128 kbps"),
-InternetRadioStation("fm187", "Retrosounds", "https://listen.radioking.com/radio/34544/stream/70739", "Retro", "128 kbps"),
-InternetRadioStation("fm188", "Stad den Haag", "https://stream.radiostaddenhaag.com/stream/1/;", "Retro", "128 kbps"),
-InternetRadioStation("fm189", "TopSide", "https://listen.radioking.com/radio/33560/stream/69731", "Retro", "128 kbps"),
-InternetRadioStation("fm190", "Affinity", "https://stream.radio.co/s038d07e75/listen", "Retro", "128 kbps"),
-InternetRadioStation("fm191", "TrendFM Den Haag", "https://trendfm.live-streams.nl:18030/live", "Retro", "128 kbps"),
-InternetRadioStation("fm192", "neoFM", "https://streams.fluxfm.de/neofm/mp3-128/streams.fluxfm.de", "Classical", "128 kbps"),
-InternetRadioStation("fm193", "Cinemix", "https://kathy.torontocast.com:1825/stream", "Classical", "128 kbps"),
-InternetRadioStation("fm194", "Minnesota Public Radio", "https://cms.stream.publicradio.org/cms.aac", "Classical", "128 kbps"),
-InternetRadioStation("fm195", "Stephansdom", "https://radioklassikstephansdom.ice.infomaniak.ch/radioklassikstephansdom.mp3", "Classical", "128 kbps"),
-InternetRadioStation("fm196", "Bru Zane", "https://uk2.streamingpulse.com/ssl/bzr", "Classical", "128 kbps"),
-InternetRadioStation("fm197", "Ancientfm", "https://mediaserv73.live-streams.nl:18058/stream", "Classical", "128 kbps"),
-InternetRadioStation("fm198", "Swiss Classic", "https://stream.srg-ssr.ch/m/rsc_de/mp3_128", "Classical", "128 kbps"),
-InternetRadioStation("fm199", "Venice classic", "https://uk2.streamingpulse.com/ssl/vcr1", "Classical", "128 kbps"),
-InternetRadioStation("fm200", "Radio Classique", "https://radioclassique.ice.infomaniak.ch/radioclassique-high.mp3", "Classical", "128 kbps"),
-InternetRadioStation("fm201", "Klassik Radio", "https://stream.klassikradio.de/klr-krdnational/aac-128/konsoleplayer", "Classical", "128 kbps"),
-InternetRadioStation("fm202", "RCO Live", "https://ice.cr6.streamzilla.xlcdn.com:8000/sz=RCOLiveWebradio=mp3-192", "Classical", "128 kbps"),
-InternetRadioStation("fm203", "Accent 4", "https://str0.creacast.com/accent4", "Classical", "128 kbps"),
-InternetRadioStation("fm204", "Schwarzenstein", "https://streams.fluxfm.de/jazzschwarz/mp3-128/streams.fluxfm.de", "Jazz", "128 kbps"),
-InternetRadioStation("fm205", "Concertzender", "https://streams.greenhost.nl:8006/live", "Jazz", "128 kbps"),
-InternetRadioStation("fm206", "Giants of Jazz", "https://streaming.radio.co/s297e618a7/listen", "Jazz", "128 kbps"),
-InternetRadioStation("fm207", "Megaton Cafe", "https://us2.internet-radio.com/proxy/megatoncafe?mp=/stream;", "Jazz", "128 kbps"),
-InternetRadioStation("fm208", "Swing Street", "https://s1.voscast.com:10413/stream", "Jazz", "128 kbps"),
-InternetRadioStation("fm209", "Couleurs Jazz", "https://listen.radioking.com/radio/127546/stream/167344", "Jazz", "128 kbps"),
-InternetRadioStation("fm210", "Smoothjazz", "https://smoothjazz.cdnstream1.com/2585_128.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm211", "Radio Swiss Jazz", "https://stream.srg-ssr.ch/m/rsj/mp3_128", "Jazz", "128 kbps"),
-InternetRadioStation("fm212", "Smooth Jazz Florida", "https://usa6.fastcast4u.com/proxy/wsjfhd?mp=/1", "Jazz", "128 kbps"),
-InternetRadioStation("fm213", "WRTI", "https://wrti-live.streamguys1.com/jazz-mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm214", "TSF 89.9", "https://tsfjazz.ice.infomaniak.ch/tsfjazz-high.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm215", "Jazzradio", "https://jazzradio.ice.infomaniak.ch/jazzradio-high.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm216", "QCIndie", "https://stream.qcindie.com/live", "Rock", "128 kbps"),
-InternetRadioStation("fm217", "Loaded", "https://s2.radio.co/sad782b7f8/listen", "Rock", "128 kbps"),
-InternetRadioStation("fm218", "XRaydio", "https://xraydio.ddns.net/live", "Rock", "128 kbps"),
-InternetRadioStation("fm219", "Cathedral 13", "https://usa15.fastcast4u.com/proxy/tstamate?mp=/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm220", "ZZROCK", "https://axtran.radioca.st/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm221", "Wigwam", "https://streaming.radio.co/s2b39cfe62/listen", "Rock", "128 kbps"),
-InternetRadioStation("fm222", "Spoon", "https://spoonradio.ice.infomaniak.ch/spoonradio-light.aac", "Rock", "128 kbps"),
-InternetRadioStation("fm223", "SA Rock", "https://sarock.radioca.st/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm224", "DKFM Shoegaze", "https://kathy.torontocast.com:2005/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm225", "Cosmic Fuzz", "https://15113.live.streamtheworld.com/SAM04AAC335_SC", "Rock", "128 kbps"),
-InternetRadioStation("fm226", "Caroline", "https://stream.radiocaroline.net/;*.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm227", "Surf Rock", "https://radio.dripfeed.net/radio/8020/radio.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm228", "Krautrock world", "https://krautrockworld.stream.laut.fm/krautrockworld", "Rock", "128 kbps"),
-InternetRadioStation("fm229", "Morow", "https://stream.fr.morow.com/morow_med.aacp", "Rock", "128 kbps"),
-InternetRadioStation("fm230", "Bluesradio", "https://i4.streams.ovh/sc/bluesrad/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm231", "Delicious Agony", "https://deliciousagony.streamguys1.com", "Rock", "128 kbps"),
-InternetRadioStation("fm232", "Rock Melodic", "https://i4.streams.ovh:2200/ssl/rockmelo?mp=/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm233", "OUI FM", "https://ouifm.ice.infomaniak.ch/ouifm-high.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm234", "Idobi Howl", "https://idobihowl.idobi.com/;stream.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm235", "12XU", "https://12xu.stream.laut.fm/12xu", "Rock", "128 kbps"),
-InternetRadioStation("fm236", "Free Phoenix", "https://s6.reliastream.com/proxy/radiofre?mp=/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm237", "La Grosse Radio Rock", "https://hd.lagrosseradio.info/lagrosseradio-rock-192.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm238", "Paul in Rio", "https://usa10.fastcast4u.com/paulinrio/;*.mp3", "World", "128 kbps"),
-InternetRadioStation("fm239", "Exotica", "https://solid67.streamupsolutions.com/proxy/repadnom?mp=/;type=mp3", "World", "128 kbps"),
-InternetRadioStation("fm240", "ILoveRio", "https://eu10.fastcast4u.com/iloverio", "World", "128 kbps"),
-InternetRadioStation("fm241", "Salseo", "https://stream.radio.co/s579ede15c/listen", "World", "128 kbps"),
-            
-            // ==================================================================
-            // streamdir.com Collection (190 stations)
-            // Source: https://www.streamdir.com/ (filtered for quality)
-            // ==================================================================
-            InternetRadioStation("sd2", "Nonstop exclusive DJ-Mixes in all styles of Trance", "http://www.tranceradio.ch/listen.m3u", "Trance", "128 kbps"),
-            InternetRadioStation("sd3", "Italian music radio of years 60 and 70. Only beaut", "http://159.8.16.16:8030/listen.pls", "60s", "128 kbps"),
-            InternetRadioStation("sd4", "Revoradio 104.1 Fm Cabot Arkansas best new rock Al", "http://174.122.26.41:80/listen.pls", "Alternative", "128 kbps"),
-            InternetRadioStation("sd5", "Revoradio 104.1 Fm Cabot Arkansas Best New Rock Al", "http://69.175.13.130:8220/listen.pls", "Alternative", "192 kbps"),
-            InternetRadioStation("sd6", "DCR - Bringing you the best of the bass driven 24/", "http://173.192.48.97:8343/listen.pls", "Drum 'n' Bass", "128 kbps"),
-            InternetRadioStation("sd7", "Online Rock Radio and Music Site streaming Alterna", "http://213.239.206.179:2199/tunein/rockxs.pls", "Rock", "128 kbps"),
-            InternetRadioStation("sd10", ". radio xenu .", "http://184.95.32.92:8000/relay.m3u", "Alternative", "96 kbps"),
-            InternetRadioStation("sd11", "fnoob techno", "http://109.108.129.75:8092/listen.pls", "Dance", "192 kbps"),
-            InternetRadioStation("sd12", "fnoob underground", "http://109.108.129.75:8090/listen.pls", "Electronic", "192 kbps"),
-            InternetRadioStation("sd13", "Radio Gibson Music", "http://ascolta.radiogibson.net:8000/listen.pls", "Pop", "128 kbps"),
-            InternetRadioStation("sd14", "SPARKS.FM UK RADIO", "http://37.187.93.104:8299/listen.pls", "Drum 'n' Bass", "128 kbps"),
-            InternetRadioStation("sd15", "! GayRadioBec.com !", "http://87.98.172.61:8000/listen.pls", "Pop", "192 kbps"),
-            InternetRadioStation("sd17", "! Sound FM 88 - The chillout lounge !", "http://ares.sxcore.net:9330/listen.pls", "Electronic", "96 kbps"),
-            InternetRadioStation("sd19", "!* Sound FM 88 - chillout & lounge", "http://stream5.sxcore.net:9820/listen.pls", "Electronic", "128 kbps"),
-            InternetRadioStation("sd20", "#1 Hip Hop Radio", "http://hearme.fm:2199/tunein/i_1hiphopradio.pls", "Hip-Hop/Rap", "128 kbps"),
-            InternetRadioStation("sd21", "#HipHop", "http://hearme.fm:2199/tunein/i_hiphop-2.pls", "Hip-Hop/Rap", "128 kbps"),
-            InternetRadioStation("sd22", "%u0160tivor", "http://163.172.213.155:8091/listen.pls", "Rock", "96 kbps"),
-            InternetRadioStation("sd23", "((( UKBounceOne )))", "http://173.193.223.145:8062/listen.pls", "Pop", "128 kbps"),
-            InternetRadioStation("sd25", "*z100", "http://88.191.74.188:7500/listen.pls", "Pop", "128 kbps"),
-            InternetRadioStation("sd26", "- Max Mashup Megamix -", "http://109.74.196.48/stream/djleon.pls", "Dance", "128 kbps"),
-            InternetRadioStation("sd27", "- Radio Gibson Music -", "http://ascolta.radiogibson.net/radiogibsonaac.m3u", "Pop", "128 kbps"),
-            InternetRadioStation("sd30", ". JellyRadio.com", "http://sc128.jellyradio.com:8000/listen.pls", "Hip-Hop/Rap", "128 kbps"),
-            InternetRadioStation("sd32", ". JellyRadio.com", "http://sc192.jellyradio.com:8004/listen.pls", "Hip-Hop/Rap", "192 kbps"),
-            InternetRadioStation("sd36", "1 Stop Hip Hop", "http://hearme.fm:2199/tunein/i_1stophiphop.pls", "Hip-Hop/Rap", "128 kbps"),
-            InternetRadioStation("sd38", "10 Years of Dance", "http://hearme.fm:2199/tunein/i_10yearsofdance.pls", "Dance", "128 kbps"),
-            InternetRadioStation("sd41", "107.5 andHow.FM", "http://184.154.90.186:8123/listen.pls", "Indie", "128 kbps"),
-            InternetRadioStation("sd42", "107.5 Radio TransAmerica", "http://208.43.9.96:8282/listen.pls", "Techno", "192 kbps"),
-            InternetRadioStation("sd43", "108.8 The End - Canada\'s Alternative", "http://149.56.147.197:8330/listen.pls", "Alternative", "128 kbps"),
-            InternetRadioStation("sd45", "123 Dance", "http://hearme.fm:2199/tunein/i_123dance.pls", "Dance", "128 kbps"),
-            InternetRadioStation("sd46", "181.fm", "http://205.188.215.230:8000/listen.pls", "70s", "128 kbps"),
-            InternetRadioStation("sd47", "1940s Radio Station", "http://199.189.111.28:8012/listen.pls", "Jazz", "128 kbps"),
-            InternetRadioStation("sd48", "1a Schlager", "http://1aradio.radionetz.de:8000/1a-schlager.mp3.m3u", "World", "128 kbps"),
-            InternetRadioStation("sd49", "1a Schlager", "http://1aradio.radionetz.de:8000/1a-schlager.mp3.m3u", "World", "128 kbps"),
-            InternetRadioStation("sd50", "1Mix Radio", "http://87.98.173.107:8010/listen.pls", "House", "192 kbps"),
-            InternetRadioStation("sd51", "1Mix Radio", "http://www.1mix.co.uk/listenh.m3u", "Trance", "192 kbps"),
-            InternetRadioStation("sd52", "1Mix Radio", "http://87.98.173.107:8000/listen.pls", "Trance", "192 kbps"),
-            InternetRadioStation("sd54", "2 loco radio usa", "https://ca2.radioboss.fm:2199/tunein/rabarton.pls", "Rock", "320 kbps"),
-            InternetRadioStation("sd55", "20 Years of House", "http://hearme.fm:2199/tunein/i_20yearsofhouse.pls", "House", "128 kbps"),
-            InternetRadioStation("sd57", "21st Century Dance", "http://hearme.fm:2199/tunein/r_21stcenturydance.pls", "Dance", "128 kbps"),
-            InternetRadioStation("sd60", "24/7 House", "http://hearme.fm:2199/tunein/i_247house.pls", "House", "128 kbps"),
-            InternetRadioStation("sd62", "247 the van", "http://192.99.17.12:5880/listen.pls", "70s", "96 kbps"),
-            InternetRadioStation("sd68", "3of9 The Collective", "https://control.internet-radio.com:2199/tunein/30f9.pls", "Alternative", "128 kbps"),
-            InternetRadioStation("sd71", "420 Fm - The Stoner Radio Network", "http://voice.420fm.cf:8000/listen.pls", "60s", "128 kbps"),
-            InternetRadioStation("sd72", "42Radio", "http://142.4.217.133:9797/listen.pls", "Rock", "128 kbps"),
-            InternetRadioStation("sd76", "60s CHARTBUSTERS", "http://37.187.90.121:3242/listen.pls", "60s", "96 kbps"),
-            InternetRadioStation("sd77", "64.150.176.42", "http://64.150.176.42:8263/listen.pls", "Jazz", "128 kbps"),
-            InternetRadioStation("sd79", "70.36.96.112", "http://70.36.96.112:8362/listen.pls", "80s", "128 kbps"),
-            InternetRadioStation("sd80", "70s 80s All Time Greatest...", "http://94.23.62.189:5500/listen.pls", "70s", "96 kbps"),
-            InternetRadioStation("sd81", "70s CHARTBUSTERS", "http://192.99.4.210:3132/listen.pls", "70s", "96 kbps"),
-            InternetRadioStation("sd82", "70vibe-FM", "http://162.244.80.21:9516/listen.pls", "70s", "128 kbps"),
-            InternetRadioStation("sd83", "80 Remember Sounds", "http://80remembersounds.radio2listen.com/listen.pls", "80s", "128 kbps"),
-            InternetRadioStation("sd85", "80s CHARTBUSTERS", "http://192.99.4.210:3130/listen.pls", "80s", "96 kbps"),
-            InternetRadioStation("sd86", "80s Music Station", "http://51.83.143.203:8580/listen.pls", "80s", "192 kbps"),
-            InternetRadioStation("sd87", "888RADIO", "http://listen.radionomy.com/888radio.m3u", "60s", "128 kbps"),
-            InternetRadioStation("sd88", "90s CHARTBUSTERS", "http://192.99.4.210:3406/listen.pls", "90s", "96 kbps"),
-            InternetRadioStation("sd89", ":: ELF 2 ::", "http://sc.ipip.cz:8012/listen.pls", "World", "128 kbps"),
-            InternetRadioStation("sd90", "a*play", "http://s12.myradiostream.com/14832.pls", "Indie", "128 kbps"),
-            InternetRadioStation("sd95", "Aaron FM (80\'s)", "http://cent5.serverhostingcenter.com/tunein.php/action200/playlist.pls", "80s", "128 kbps"),
-            InternetRadioStation("sd96", "Aaron FM (90s Hits)", "http://radio.shoutcastbroadcaster.com/tunein.php/acanada1/playlist.pls", "90s", "128 kbps"),
-            InternetRadioStation("sd98", "Aaron FM (Country))", "http://radio.shoutcastbroadcaster.com/tunein.php/acanada3/playlist.pls", "Country", "128 kbps"),
-            InternetRadioStation("sd108", "Absolute Tunage", "http://cp.internet-radio.org.uk/tunein.php/absolutetunage/playlist.pls", "Trance", "128 kbps"),
-            InternetRadioStation("sd110", "AbsoluteAnthems", "http://cast.iplayradio.net:2199/tunein/anthems2.pls", "Electronic", "128 kbps"),
-            InternetRadioStation("sd114", "Access Heavens Country", "http://hearme.fm:2199/tunein/r_heavenscountry.pls", "Country", "128 kbps"),
-            InternetRadioStation("sd115", "Acidic Infektion Radio", "http://173.193.195.155:9976/listen.pls", "Industrial", "128 kbps"),
-            InternetRadioStation("sd116", "Actone Jazz", "http://hearme.fm:2199/tunein/r_actonejazz.pls", "Jazz", "128 kbps"),
-            InternetRadioStation("sd118", "Affinity Radio", "http://50.7.242.114:7662/listen.pls", "60s", "128 kbps"),
-            InternetRadioStation("sd122", "African FM", "http://africaninternetradio.com:8000/listen.pls", "World", "128 kbps"),
-            InternetRadioStation("sd123", "African hitz radio", "http://www.centovav3.shoutstream.co.uk/tunein/daarthur.pls", "World", "128 kbps"),
-            InternetRadioStation("sd126", "AfterHoursDJ", "http://www.afterhoursdjs.org/ahdjs128.m3u", "Trance", "128 kbps"),
-            InternetRadioStation("sd127", "afterlife radio", "http://cast.iplayradio.net:8069/listen.pls", "Electronic", "320 kbps"),
-            InternetRadioStation("sd139", "All Southern Gospel Radio", "http://www.icastcenter.com/cast/tunein.php/marglass/playlist.pls", "Country", "128 kbps"),
-            InternetRadioStation("sd144", "Always Country", "http://208.85.240.89:80/listen.pls", "Country", "96 kbps"),
-            InternetRadioStation("sd145", "Amazing Smooth and Jazz", "http://yp.shoutcast.com/sbin/tunein-station.pls?id=178027&play_status=1", "Jazz", "96 kbps"),
-            InternetRadioStation("sd147", "Ambient Art Foundation", "http://188.26.103.151:8090/listen.pls", "Ambient", "128 kbps"),
-            InternetRadioStation("sd148", "Ambient Art Foundation", "http://185.80.220.12:8154/listen.pls", "Ambient", "128 kbps"),
-            InternetRadioStation("sd149", "Ambient Art Foundation", "http://185.80.220.12:8062/listen.pls", "Ambient", "128 kbps"),
-            InternetRadioStation("sd150", "Ambient Lounge on MixLive.ie", "http://198.24.145.146:9332/listen.pls", "Ambient", "128 kbps"),
-            InternetRadioStation("sd151", "ambient radio -anotherplanet.fm -", "http://cast1.serverhostingcenter.com/tunein.php/evarakin/playlist.pls", "Ambient", "128 kbps"),
-            InternetRadioStation("sd155", "American road Radio", "http://user.radioboss.fm:2199/tunein/na00.pls", "Rock", "128 kbps"),
-            InternetRadioStation("sd156", "Amnesia", "http://hearme.fm:2199/tunein/i_amnesia.pls", "House", "128 kbps"),
-            InternetRadioStation("sd160", "Andrew Lord", "http://hearme.fm:2199/tunein/andrewlord.pls", "House", "128 kbps"),
-            InternetRadioStation("sd165", "Anime Classics Radio", "http://192.99.8.170:9165/listen.pls", "Asian", "128 kbps"),
-            InternetRadioStation("sd166", "anime2u", "http://176.31.107.8:8254/listen.pls", "J-Music", "128 kbps"),
-            InternetRadioStation("sd175", "Armageddeonmusik", "http://cent5.serverhostingcenter.com:2199/tunein/hpgrwiqo.pls", "Reggae", "128 kbps"),
-            InternetRadioStation("sd180", "Array.FM Country", "http://75.126.64.105:8187/listen.pls", "Country", "128 kbps"),
-            InternetRadioStation("sd186", "Array.FM Rhythmic", "http://75.126.64.105:8103/listen.pls", "R'n'B", "128 kbps"),
-            InternetRadioStation("sd199", "B2BGospelRadio", "http://108.163.210.108:80/listen.pls", "R'n'B", "128 kbps"),
-            InternetRadioStation("sd201", "Babaganousha Radio", "http://91.121.143.55:8000/listen.pls", "Psychedelic", "96 kbps"),
-            InternetRadioStation("sd202", "Babaganousha Radio", "http://188.68.42.236:8000/listen.pls", "Psychedelic", "128 kbps"),
-            InternetRadioStation("sd207", "Bass Beats Radio", "http://174.37.16.73:1715/listen.pls", "Drum 'n' Bass", "128 kbps"),
-            InternetRadioStation("sd211", "Bassdrive", "http://hearme.fm:2199/tunein/r_bassdrive.pls", "Drum 'n' Bass", "128 kbps"),
-            InternetRadioStation("sd212", "BassDrive", "http://www.bassdrive.com/v2/streams/BassDrive.m3u", "Drum 'n' Bass", "128 kbps"),
-            InternetRadioStation("sd226", "Beethoven Channel", "http://82.94.205.106:8021/listen.pls", "Classical", "320 kbps"),
-            InternetRadioStation("sd238", "Black Aces radio", "http://knight.wavestreamer.com:2151/listen.m3u?sid=1", "Indie", "128 kbps"),
-            InternetRadioStation("sd241", "BLESS FM", "http://178.159.0.13:8086/listen.pls", "Reggae", "128 kbps"),
-            InternetRadioStation("sd245", "BluesMen Channel (Gold)", "https://bluesmenchannel.com/Radio/Radio-Playlist/Gold/WinampiTunes/BluesMen_Channel_Gold-WinampiTunes.m3u", "Blues", "320 kbps"),
-            InternetRadioStation("sd246", "BluesMen Channel (Hits)", "https://bluesmenchannel.com/Radio/Radio-Playlist/Hits/WinampiTunes/BluesMen_Channel_Hits-WinampiTunes.m3u", "Blues", "320 kbps"),
-            InternetRadioStation("sd247", "BluesMen Channel - Gold", "https://bluesmenchannel.com/Radio/Radio-Playlist/Hits/WinampiTunes/BluesMen_Channel-Hits-Radio_Playlist_WinampiTunes-.pls", "Blues", "320 kbps"),
-            InternetRadioStation("sd248", "BluesMen Channel - Hits", "https://bluesmenchannel.com/Radio/Radio-Playlist/Hits/WinampiTunes/BluesMen_Channel-Hits-Radio_Playlist_WinampiTunes-.pls", "Blues", "320 kbps"),
-            InternetRadioStation("sd249", "BlueTrain Classics - The Jazz UK", "http://46.28.49.165/stream/bluetrain.pls?mp=/", "Jazz", "128 kbps"),
-            InternetRadioStation("sd251", "Boneyaad Radio", "http://69.175.103.226:8180/listen.pls", "Reggae", "96 kbps"),
-            InternetRadioStation("sd255", "Braingell Radio", "http://hearme.fm:2199/tunein/r_braingell.pls", "Metal", "128 kbps"),
-            InternetRadioStation("sd261", "Broseleyfm", "https://hubble.shoutca.st/tunein/agriffit.pls", "Indie", "128 kbps"),
-            InternetRadioStation("sd262", "Brutal Existence Radio", "http://brutalexistenceradio.com:8080/listen.pls", "Metal", "128 kbps"),
-            InternetRadioStation("sd263", "btz radio", "http://stereo.wavestreamer.com:9019/listen.pls?sid=1", "Dubstep", "128 kbps"),
-            InternetRadioStation("sd267", "Cali Smoke Radio", "http://198.178.123.5:8156/listen.pls", "Reggae", "128 kbps"),
-            InternetRadioStation("sd272", "Carnavaleando Radio", "http://198.58.106.133/stream/carnavaleando.pls?mp=", "Latin", "128 kbps"),
-            InternetRadioStation("sd281", "Chart Fm", "http://cast.iplayradio.net:2199/tunein/chartfm.pls", "90s", "128 kbps"),
-            InternetRadioStation("sd285", "Chile Continental", "http://69.57.166.8:8029/listen.pls", "90s", "128 kbps"),
-            InternetRadioStation("sd291", "CINZ NET Radio", "https://stream.zeno.fm/6cyjqwlkz4ptv.pls", "J-Music", "128 kbps"),
-            InternetRadioStation("sd298", "City Radio Preston", "http://87.117.228.94:7220/listen.pls", "R'n'B", "128 kbps"),
-            InternetRadioStation("sd308", "Clubberz-Paradise", "http://50.30.32.210:8134/listen.pls", "Hardcore", "96 kbps"),
-            InternetRadioStation("sd309", "Clube Brasil", "http://www.clubebrasil.fr/winamphq.pls", "Latin", "192 kbps"),
-            InternetRadioStation("sd310", "Clubsoundz", "http://62.210.107.52:8020/listen.pls", "Techno", "320 kbps"),
-            InternetRadioStation("sd312", "Colombia Radial", "http://144.217.195.24:8597/listen.pls", "Latin", "320 kbps"),
-            InternetRadioStation("sd313", "Colombo Radio", "http://5.39.71.159:8357/listen.pls", "Classical", "320 kbps"),
-            InternetRadioStation("sd316", "CoolFm Hits", "http://144.217.129.213:8938/listen.pls", "90s", "128 kbps"),
-            InternetRadioStation("sd317", "Cornucopia Broadcasting", "http://orange.citrus3.com:8212/listen.pls", "Experimental", "128 kbps"),
-            InternetRadioStation("sd340", "Dance Uk", "http://83.170.104.91:30252/listen.pls", "Hardcore", "128 kbps"),
-            InternetRadioStation("sd355", "Dark Soul 7", "http://hearme.fm:2199/tunein/r_darksoul7.pls", "Metal", "128 kbps"),
-            InternetRadioStation("sd358", "Datablender.fm", "http://hearme.fm:2199/tunein/r_datablenderfm.pls", "Techno", "128 kbps"),
-            InternetRadioStation("sd363", "Dead To Self Radio", "http://198.154.112.233:8613/listen.pls", "Hardcore", "128 kbps"),
-            InternetRadioStation("sd376", "Depressive metal rock Radio", "http://184.154.10.83:8390/listen.pls", "Metal", "128 kbps"),
-            InternetRadioStation("sd382", "DESIFM &JATTFM  punjabi  bhangra  hindi urban desi", "http://s30.myradiostream.com:18208/listen.pls", "Asian", "128 kbps"),
-            InternetRadioStation("sd403", "DJKhaos Metal Show", "http://184.173.181.32:36342//listen.mp3.m3u", "Metal", "128 kbps"),
-            InternetRadioStation("sd426", "Dude Suit Radio", "http://www.dude-suit.net:8000/listen.pls", "Trip Hop", "128 kbps"),
-            InternetRadioStation("sd427", "Dude Suit Radio", "http://www.dude-suit.net:8000/live.mp3.m3u", "Trip Hop", "128 kbps"),
-            InternetRadioStation("sd438", "EDMONTUNES", "http://176.31.111.65:9104/listen.pls", "Indie", "192 kbps"),
-            InternetRadioStation("sd439", "EL MEGAHUEV", "http://192.99.8.170:9897/listen.pls", "Latin", "128 kbps"),
-            InternetRadioStation("sd447", "Elim Radio UK", "http://cent4.serverhostingcenter.com:2199/tunein/jjrklwdq.pls", "Classical", "128 kbps"),
-            InternetRadioStation("sd468", "Estereo Uno", "http://66.70.187.44:9189/listen.pls", "Latin", "128 kbps"),
-            InternetRadioStation("sd475", "Exquisite Soul Radio", "http://37.187.90.196:8896/listen.pls", "R'n'B", "128 kbps"),
-            InternetRadioStation("sd476", "Facets of Art Radio", "http://109.74.196.76:8157//stream.m3u", "Experimental", "320 kbps"),
-            InternetRadioStation("sd492", "Folk Tyme [RadioAvenue.com]", "http://209.95.50.189/stream/johenger.pls", "Folk", "128 kbps"),
-            InternetRadioStation("sd493", "Folk Tyme [RadioAvenue.com]", "http://69.162.125.2:2199/tunein/rafolk.pls", "Folk", "128 kbps"),
-            InternetRadioStation("sd514", "Friese Piraten", "http://server1.sfm-stream.nl:9000/listen.pls?sid=1", "Folk", "192 kbps"),
-            InternetRadioStation("sd516", "Fuel Radio", "http://209.105.250.69:8479/listen.pls", "Hardcore", "128 kbps"),
-            InternetRadioStation("sd520", "Funky Corner Radio", "https://andromeda.shoutca.st/tunein/funkycor.pls", "Funk", "128 kbps"),
-            InternetRadioStation("sd542", "GlassOnion Radio", "http://www.streamlicensing.com/directory/index.cgi/playlist.pls?action=playlist&type=pls&sid=4593&stream_id=8666", "Blues", "128 kbps"),
-            InternetRadioStation("sd547", "Go Punk Yourself Radio", "http://204.45.29.98:4276/listen.pls", "Punk", "128 kbps"),
-            InternetRadioStation("sd552", "Gospel Truth Lessons", "http://198.245.50.150:6826/listen.pls", "Soul", "192 kbps"),
-            InternetRadioStation("sd561", "greenfields online radio", "http://uk2.internet-radio.com/tunein.php/emeraldgold/playlist.pls", "Folk", "128 kbps"),
-            InternetRadioStation("sd569", "GTriddim Guyana Radio", "http://172.246.122.146:8080/listen.pls", "Reggae", "128 kbps"),
-            InternetRadioStation("sd576", "HARDER-FM THE HARDERSOUND", "http://harder-fm.com:8000/listen.pls?sid=1", "Hardcore", "192 kbps"),
-            InternetRadioStation("sd602", "Hot 88 FM 88.8 WTET", "http://fastcast4u.com:9574/listen.pls", "R'n'B", "128 kbps"),
-            InternetRadioStation("sd623", "Hussieskunk Punk Radio", "http://69.175.13.131:8110/listen.pls", "Punk", "128 kbps"),
-            InternetRadioStation("sd664", "Josh Holiday", "http://hearme.fm:2199/tunein/hmjoshholiday.pls", "Techno", "128 kbps"),
-            InternetRadioStation("sd666", "JPopsuki Radio!", "http://jpopsuki.fm:2199/tunein/jpopsuki-stream.pls", "J-Music", "192 kbps"),
-            InternetRadioStation("sd675", "JUST VIBE RADIO", "http://37.26.106.12:9998/listen.pls", "Garage", "128 kbps"),
-            InternetRadioStation("sd676", "Just Vibe radio", "http://www.justvibe.co.uk/listen.pls", "Garage", "128 kbps"),
-            InternetRadioStation("sd681", "Kavi FM Tamil", "http://93.190.141.15:7067/listen.pls", "Asian", "128 kbps"),
-            InternetRadioStation("sd692", "Kiss LGBT Radio", "https://nebula.shoutca.st/tunein/tamaci02.pls", "Techno", "128 kbps"),
-            InternetRadioStation("sd704", "KRUSH LIVE", "http://64.71.162.108:8715/listen.pls", "Garage", "128 kbps"),
-            InternetRadioStation("sd706", "KTLK-DB The Fringe FM", "http://psn3.prostreaming.net:2199/tunein/joserupe.pls", "Experimental", "128 kbps"),
-            InternetRadioStation("sd724", "limbikfreq", "http://205.188.215.225:8018/listen.pls", "IDM", "128 kbps"),
-            InternetRadioStation("sd736", "LoveFire.ca Radio - Music, Talk, Alternative News ", "http://streams2.museter.com/tunein.php/lovefire/tunein.pls", "Dubstep", "128 kbps"),
-            InternetRadioStation("sd745", "Maerong SNSDCentral Radio", "http://216.59.35.34:8088/listen.pls", "Asian", "128 kbps"),
-            InternetRadioStation("sd790", "More Fm - Classic Rock Radio", "http://216.66.84.2:6320/listen.pls", "Post-Rock", "128 kbps"),
-            InternetRadioStation("sd827", "New Orleans Radio", "http://65.49.39.138:8002/listen.pls", "Funk", "128 kbps"),
-            InternetRadioStation("sd831", "NiaRadioNetwork", "http://www.niaradionetwork.com/gospel.m3u", "Soul", "96 kbps"),
-            InternetRadioStation("sd840", "Noise FM", "http://5.39.93.200:8000/play_256.m3u", "Dubstep", "224 kbps"),
-            InternetRadioStation("sd841", "Noise Vandals", "http://69.4.230.79:8100/listen.pls", "UK Garage", "128 kbps"),
-            InternetRadioStation("sd850", "NSB Radio", "http://www.nsbradio.co.uk/listen96k.pls", "Breaks", "96 kbps"),
-            InternetRadioStation("sd877", "Open Broadcast", "http://77.59.245.12:80/openbroadcast.m3u", "Experimental", "192 kbps"),
-            InternetRadioStation("sd879", "ORANG SUKA AKU SUKA FM", "http://myradiostream.com/21/11158.m3u", "Asian", "96 kbps"),
-            InternetRadioStation("sd891", "Parole Delivrance", "http://45.79.186.124:8400/listen.pls", "Soul", "128 kbps"),
-            InternetRadioStation("sd912", "PoingFM - Renegade Radio", "http://82.94.205.115:8160//radio.m3u", "Breaks", "320 kbps"),
-            InternetRadioStation("sd914", "Polka Jammer Network", "http://74.115.212.124:80/listen.pls", "Folk", "128 kbps"),
-            InternetRadioStation("sd916", "POSTROCKMUSIC", "http://46.105.109.142:8014/listen.pls", "Post-Rock", "128 kbps"),
-            InternetRadioStation("sd917", "povradio", "http://173.192.224.123:8600/listen.pls", "Dubstep", "128 kbps"),
-            InternetRadioStation("sd931", "Psychedelic FM", "http://www.psychedelicfm.com/listen.pls", "Psychedelic", "128 kbps"),
-            InternetRadioStation("sd1026", "Radio Riel - The Brideswell Stream", "http://66.186.45.115:9012//live.m3u", "Classical", "96 kbps"),
-            InternetRadioStation("sd1032", "Radio Schizoid - Psychedelic Trance/Goa Trance/Pro", "http://schizoid.in/schizoid-psy.pls", "Psychedelic", "128 kbps"),
-            InternetRadioStation("sd1070", "radiOi", "http://85.17.30.90:8090/listen.pls", "Punk", "96 kbps"),
-            InternetRadioStation("sd1099", "Real Punk Radio", "http://69.175.60.162/listen.pls", "Punk", "128 kbps"),
-            InternetRadioStation("sd1127", "RJM Soul", "http://31.12.64.24:8000/jmaster.m3u", "Soul", "128 kbps"),
-            InternetRadioStation("sd1153", "salalondon", "http://cp6.shoutcheap.com:2199/tunein/smartins.pls", "Funk", "96 kbps"),
-            InternetRadioStation("sd1175", "Shirley and Spinoza Radio", "http://207.200.96.227:8004/listen.pls", "Experimental", "128 kbps"),
-            InternetRadioStation("sd1218", "spaces classical radio redefined", "http://shoutcast.reliableweb.co.uk/tunein.php/latty/playlist.pls", "Classical", "128 kbps"),
-            InternetRadioStation("sd1227", "Sradiouk", "http://139.162.245.57/stream/sradiouk.pls?mp=/stream", "UK Garage", "128 kbps"),
-            InternetRadioStation("sd1253", "Sweet Sounz", "http://109.123.114.172:7333/listen.pls", "Funk", "128 kbps"),
-            InternetRadioStation("sd1254", "Sweet Sounz", "http://soho.wavestreamer.com:7333/listen.m3u?sid=1", "Soul", "128 kbps"),
-            InternetRadioStation("sd1281", "The Bluegrass Mix", "http://s4.nexuscast.com/tunein/bluegrassmix128.pls", "Bluegrass", "128 kbps"),
-            InternetRadioStation("sd1307", "The Mix Rock", "http://67.212.174.226:10011/listen.pls", "Post-Rock", "128 kbps"),
-            InternetRadioStation("sd1327", "Third Wave Ska", "http://knight.wavestreamer.com:9097/listen.m3u?sid=1", "Ska", "128 kbps"),
-            InternetRadioStation("sd1342", "Touhou   Game   Anime Radio", "http://68.68.28.167:8506/stream.m3u", "J-Music", "96 kbps"),
-            InternetRadioStation("sd1345", "Toxic Sickness Radio - Industrial-Hardcore, Hardco", "http://uk2.internet-radio.com/tunein.php/toxicsickness/playlist.pls", "Industrial", "128 kbps"),
-            InternetRadioStation("sd1386", "UKG247.com", "http://87.117.201.160:8200/listen.pls", "UK Garage", "128 kbps"),
-            InternetRadioStation("sd1387", "ukobsession.fm", "http://108.166.161.210:8991/listen.pls", "Dubstep", "128 kbps"),
-            InternetRadioStation("sd1390", "Undercast", "http://s17.myradiostream.com/8562.pls", "Punk", "96 kbps"),
-            
-            // ==================================================================
-            // radioregistry.com Collection (80 stations)
-            // Source: https://radioregistry.com/ (from 51,339 total stations)
-            // ==================================================================
-            InternetRadioStation("rr1", "_a_u_r_a_t_o_n_", "https://auraton.stream.laut.fm/_a_u_r_a_t_o_n_", "Ambient", "128 kbps"),
-            InternetRadioStation("rr2", "Poder (Power) - WLRT", "http://162.240.215.62:9300/;?nocache=1", "Folk", "128 kbps"),
-            InternetRadioStation("rr3", "[MISC.WAV FM]", "http://streaming.radio.co/sac586d131/listen", "Electronic", "128 kbps"),
-            InternetRadioStation("rr4", "@BigBoxRadio | The BOX (WBBR-DB) HD1", "http://streaming.live365.com/a91412", "Hip Hop", "128 kbps"),
-            InternetRadioStation("rr5", "@FM (Arroba FM) - XEGDL", "http://sts.pergom.mx/xhgdl", "Latin", "128 kbps"),
-            InternetRadioStation("rr6", "+255 Global Radio", "http://zas7.ndx.co.za:18026/stream", "Pop", "128 kbps"),
-            InternetRadioStation("rr7", "~ Hit A Jam ~", "http://www.radioking.com/play/hit-a-jam", "Electronic", "128 kbps"),
-            InternetRadioStation("rr8", "0-24 80er Pop Rock Oldies", "https://stream.laut.fm/0-24_80er_pop_rock?nocache=1", "80s", "128 kbps"),
-            InternetRadioStation("rr9", "0-24 Oldies Pop Rock", "https://stream.laut.fm/0-24_oldies_pop_rock?nocache=1", "80s", "128 kbps"),
-            InternetRadioStation("rr10", "001 ROCK", "http://stream.laut.fm/-001rock", "Rock", "128 kbps"),
-            InternetRadioStation("rr11", "0100 Splash Radio", "http://streaming.live365.com/a22522", "Alternative", "128 kbps"),
-            InternetRadioStation("rr12", "011.FM - 60s", "https://listen.011fm.com/stream06?nocache=1", "60s", "128 kbps"),
-            InternetRadioStation("rr13", "011.FM - 70s", "https://listen.011fm.com/stream04?nocache=1", "70s", "128 kbps"),
-            InternetRadioStation("rr14", "011.FM - 80s", "https://listen.011fm.com/stream05?nocache=1", "80s", "128 kbps"),
-            InternetRadioStation("rr15", "011.FM - 90s", "https://listen.011fm.com/stream06?nocache=1", "90s", "128 kbps"),
-            InternetRadioStation("rr16", "011.FM - 90s Alternative", "https://listen.011fm.com/stream17?nocache=1", "Alternative", "128 kbps"),
-            InternetRadioStation("rr17", "011.FM - 90s Country", "https://listen.011fm.com/stream10?nocache=1", "Country", "128 kbps"),
-            InternetRadioStation("rr18", "011.FM - Adult Standards", "https://listen.011fm.com/stream20?nocache=1", "Jazz", "128 kbps"),
-            InternetRadioStation("rr19", "011.FM - Alternative", "https://listen.011fm.com/stream15?nocache=1", "Alternative", "128 kbps"),
-            InternetRadioStation("rr20", "011.FM - Classic Country", "https://listen.011fm.com/stream19?nocache=1", "Country", "128 kbps"),
-            InternetRadioStation("rr21", "011.FM - Classic Rock", "https://listen.011fm.com/stream02?nocache=1", "Rock", "128 kbps"),
-            InternetRadioStation("rr22", "011.FM - Country", "https://listen.011fm.com/stream11?nocache=1", "Country", "128 kbps"),
-            InternetRadioStation("rr23", "011.FM - Jazz", "https://listen.011fm.com/stream14?nocache=1", "Jazz", "128 kbps"),
-            InternetRadioStation("rr24", "011.FM - Rock", "https://listen.011fm.com/stream01?nocache=1", "Rock", "128 kbps"),
-            InternetRadioStation("rr25", "011.FM - Urban", "https://listen.011fm.com/stream25?nocache=1", "Hip Hop", "128 kbps"),
-            InternetRadioStation("rr26", "1.FM - 50s & 60s", "http://sc-1950s.1.fm:8000/listen.pls", "60s", "128 kbps"),
-            InternetRadioStation("rr27", "1.FM - 80s", "http://sc-80s.1.fm:8300/listen.pls", "80s", "128 kbps"),
-            InternetRadioStation("rr28", "1.FM - 90s", "http://sc-90s.1.fm:8040/listen.pls", "90s", "128 kbps"),
-            InternetRadioStation("rr29", "1.FM - Absolute 90s Party", "http://sc-90spartymix.1.fm/listen.pls", "90s", "128 kbps"),
-            InternetRadioStation("rr30", "1.FM - Blues", "http://sc-blues.1.fm:10240/listen.pls", "Jazz", "128 kbps"),
-            InternetRadioStation("rr31", "1.FM - Classic Country", "http://sc-classiccountry.1.fm:10160/listen.pls", "Country", "128 kbps"),
-            InternetRadioStation("rr32", "1.FM - Classic Rock", "http://sc-classicrock.1.fm:10032/listen.pls", "Rock", "128 kbps"),
-            InternetRadioStation("rr33", "1.FM - Country One", "http://sc-countryone.1.fm:8162/listen.pls", "Country", "128 kbps"),
-            InternetRadioStation("rr34", "1.FM - Disco Ball", "http://sc-disco.1.fm:8100/listen.pls", "70s", "128 kbps"),
-            InternetRadioStation("rr35", "1.FM - Jazz", "http://sc-jazz.1.fm:8032/listen.pls", "Jazz", "128 kbps"),
-            InternetRadioStation("rr36", "1.FM - Love Classics", "http://sc-loveclassics.1.fm:10080/listen.pls", "70s", "128 kbps"),
-            InternetRadioStation("rr37", "1.FM - Movie Soundtracks", "http://sc-movies.1.fm:8238/listen.pls", "Music", "128 kbps"),
-            InternetRadioStation("rr38", "1.FM - Slow Jams", "http://sc-slowjam.1.fm:10408/listen.pls", "R&B", "128 kbps"),
-            InternetRadioStation("rr39", "1.FM - Soft Rock", "http://sc-softrock.1.fm:7920/listen.pls", "Rock", "128 kbps"),
-            InternetRadioStation("rr40", "100 FM The Pike", "http://live-mp3-128.kexp.org/kexp128.mp3", "Alternative", "128 kbps"),
-            
-            // ==================================================================
-            // GitHub internet-radio-streams Collection (45 stations)
-            // Source: https://github.com/mikepierce/internet-radio-streams
-            // ==================================================================
-InternetRadioStation("fm1", "Deepershades", "https://andromeda.housejunkie.ca/radio/8000/radio.mp3", "House", "128 kbps"),
-InternetRadioStation("fm2", "Sound of Berlin", "https://fluxmusic.api.radiosphere.io/channels/sound-of-berlin/stream.aac", "House", "128 kbps"),
-InternetRadioStation("fm3", "Ibiza Global", "https://cdn-peer031.streaming-pro.com:8025/ibizaglobalradio.mp3", "House", "128 kbps"),
-InternetRadioStation("fm4", "Dogglounge", "https://securestreams4.autopo.st:1853/stream", "House", "128 kbps"),
-InternetRadioStation("fm5", "Vanilla Radio", "https://stream.vanillaradio.com:8003/stream", "House", "128 kbps"),
-InternetRadioStation("fm6", "Cavo Paradiso", "https://neos.win:48488/1", "House", "128 kbps"),
-InternetRadioStation("fm7", "Berlin Beach House Radio", "https://fluxmusic.api.radiosphere.io/channels/berlin-beach-house/stream.aac", "House", "128 kbps"),
-InternetRadioStation("fm8", "Deep House Radio", "https://streaming.shoutcast.com/dhr?1711824191948", "House", "128 kbps"),
-InternetRadioStation("fm9", "Club Sandwich", "https://streams.fluxfm.de/clubsandwich/mp3-128/streams.fluxfm.de", "House", "128 kbps"),
-InternetRadioStation("fm10", "Point Blank", "https://pointblankradio.co.uk/stream", "House", "128 kbps"),
-InternetRadioStation("fm11", "Deepinside", "https://stream.rcs.revma.com/uyrbt6xuhnruv", "House", "128 kbps"),
-InternetRadioStation("fm12", "DinamoFM Deep", "https://channels.dinamo.fm/deep-mp3", "House", "128 kbps"),
-InternetRadioStation("fm13", "Radio Deep", "https://s2.radio.co/s098f4886e/listen", "House", "128 kbps"),
-InternetRadioStation("fm14", "Ibiza Sonica Club", "https://ibizasonica.streaming-pro.com:8011/sonicaclub", "House", "128 kbps"),
-InternetRadioStation("fm15", "Mixadance", "https://stream.mixadance.fm/mixadance", "House", "128 kbps"),
-InternetRadioStation("fm16", "FG Avant Garde", "https://n11.rcs.revma.com/02s8zcv9898uv", "House", "128 kbps"),
-InternetRadioStation("fm17", "Ibiza Live Radio", "https://uksoutha.streaming.broadcast.radio/ibiza-live-radio", "House", "128 kbps"),
-InternetRadioStation("fm18", "Barcelona City fm", "https://s2.radio.co/sd9da54aa0/listen", "House", "128 kbps"),
-InternetRadioStation("fm19", "Physical", "https://s3.radio.co/sd59eccabf/listen", "House", "128 kbps"),
-InternetRadioStation("fm20", "Deep Radio", "https://29043.live.streamtheworld.com/DEEP_RADIOAAC.aac", "House", "128 kbps"),
-InternetRadioStation("fm21", "Experimental TV Radio", "https://s44.myradiostream.com/8890/listen.mp3", "House", "128 kbps"),
-InternetRadioStation("fm22", "Housefm", "https://uksoutha.streaming.broadcast.radio/housefm", "House", "128 kbps"),
-InternetRadioStation("fm23", "DinamoFM", "https://channels.dinamo.fm/legacy-mp3", "House", "128 kbps"),
-InternetRadioStation("fm24", "Frisky Deep", "https://stream.deep.friskyradio.com/mp3_low", "House", "128 kbps"),
-InternetRadioStation("fm25", "Music Factory", "https://i4.streams.ovh:2200/ssl/musicfactory?mp=/stream", "House", "128 kbps"),
-InternetRadioStation("fm26", "Fantasy FM", "https://eu4.fastcast4u.com/proxy/fantasyfm?mp=/1", "House", "128 kbps"),
-InternetRadioStation("fm27", "Deepinradio", "https://ssl1.viastreaming.net:8525/stream.mp3", "House", "128 kbps"),
-InternetRadioStation("fm28", "MoveDaHouse", "https://uk7.internet-radio.com/proxy/movedahouse?mp=//stream", "House", "128 kbps"),
-InternetRadioStation("fm29", "We get lifted", "https://s3.radio.co/sa80d22794/listen", "House", "128 kbps"),
-InternetRadioStation("fm30", "Mambo", "https://streamer.radio.co/sd7b28e5f3/listen", "House", "128 kbps"),
-InternetRadioStation("fm31", "Blue Marlin", "https://ibizasonica.streaming-pro.com:8001/bluemarlin", "House", "128 kbps"),
-InternetRadioStation("fm32", "MotionFM House", "https://vm.motionfm.com/motionone_free", "House", "128 kbps"),
-InternetRadioStation("fm33", "320fm", "https://stream.weblygo.de:8020/stream", "House", "128 kbps"),
-InternetRadioStation("fm34", "Centreforce", "https://listen.centreforceradio.com:8830/192", "House", "128 kbps"),
-InternetRadioStation("fm35", "BeachGrooves", "https://stream.beachgrooves.com:9000/;*.mp3", "House", "128 kbps"),
-InternetRadioStation("fm36", "Freak31", "https://stream.31media.net/freak31", "House", "128 kbps"),
-InternetRadioStation("fm37", "1 Radio Space ", "https://c22.radioboss.fm:18118/1RADIO.SPACE", "Techno", "128 kbps"),
-InternetRadioStation("fm38", "Frisky", "https://stream.frisky.friskyradio.com/mp3_low", "Techno", "128 kbps"),
-InternetRadioStation("fm39", "Insomnia fm", "https://insomniafm.cloudrad.io/live", "Techno", "128 kbps"),
-InternetRadioStation("fm40", "Proton radio", "https://shoutcast.protonradio.com/;*.mp3", "Techno", "128 kbps"),
-InternetRadioStation("fm41", "Joint Beat", "https://jointil.com/stream-beat", "Techno", "128 kbps"),
-InternetRadioStation("fm42", "TrancePulse", "https://stream.trance.ie:8443/stream", "Techno", "128 kbps"),
-InternetRadioStation("fm43", "Afterhours fm", "https://fr.ah.fm/live", "Techno", "128 kbps"),
-InternetRadioStation("fm44", "Trance Athena", "https://cast.streams.ovh:8008/;*.mp3", "Techno", "128 kbps"),
-InternetRadioStation("fm45", "Psyland", "https://sp.streams.ovh/8020/stream", "Techno", "128 kbps"),
-InternetRadioStation("fm46", "Discover trance", "https://uk01.discovertrance.com/;*.mp3", "Techno", "128 kbps"),
-InternetRadioStation("fm47", "FIP electro", "https://icecast.radiofrance.fr/fipelectro-midfi.mp3", "Electronic", "128 kbps"),
-InternetRadioStation("fm48", "Intergalactic", "https://radio.intergalactic.fm/1", "Electronic", "128 kbps"),
-InternetRadioStation("fm49", "Rinse UK", "https://admin.stream.rinse.fm/proxy/rinse_uk/stream", "Electronic", "128 kbps"),
-InternetRadioStation("fm50", "Sovietwave", "https://station.waveradio.org/soviet", "Electronic", "128 kbps"),
-InternetRadioStation("fm51", "Plusfm", "https://stream.plusfm.net/1", "Electronic", "128 kbps"),
-InternetRadioStation("fm52", "Biptunia", "https://ecast.myautodj.com:1380/listen.mp3", "Electronic", "128 kbps"),
-InternetRadioStation("fm53", "Noise FM", "https://play.sas-media.ru/play_256", "Electronic", "128 kbps"),
-InternetRadioStation("fm54", "Systrum Sistum Electronica", "https://systrum.net:8443/SSR2", "Electronic", "128 kbps"),
-InternetRadioStation("fm55", "Dirty", "https://live.dirty.radio/stream", "Electronic", "128 kbps"),
-InternetRadioStation("fm56", "AmpedOut", "https://s12.ssl-stream.com/8006/stream", "Electronic", "128 kbps"),
-InternetRadioStation("fm57", "Bloodlit", "https://usa5.fastcast4u.com/proxy/wwwblood?mp=/1", "Electronic", "128 kbps"),
-InternetRadioStation("fm58", "Bassdrive", "https://bassdrive.radioca.st/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm59", "Renegade", "https://streaming06.liveboxstream.uk/proxy/renegade/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm60", "Jungletrain", "https://chat.jungletrain.net/streamtest/;stream/1", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm61", "SubFM", "https://subfm.radioca.st/live", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm62", "Breakbeat Zone", "https://radio.bbz.ru/listen/bbz/192.mp3", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm63", "Dirty Mayhem dnb", "https://node-34.zeno.fm/x7d910w23d0uv", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm64", "Sunshine Live Drum-n-Bass", "https://sunsl.streamabc.net/sunsl-drumnbass-mp3-192-9834247", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm65", "UKBassRadio", "https://s2.ssl-stream.com/listen/uk_bass_radio/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm66", "Dnbradio", "https://dnbradio.nl/dnbradio_main.mp3", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm67", "Old Skool Anthemz", "https://osa.radioca.st/stream", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm68", "Laylow FM", "https://stream.laylowlondon.co.uk/listen/laylowfm/radio.ogg", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm69", "Dirtybass", "https://kentha.dirtybass.fm/endpoint/dirtybass.mp3", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm70", "Partyvibe Drum and Bass", "https://www.partyviberadio.com:8061", "Drum & Bass", "128 kbps"),
-InternetRadioStation("fm71", "WeFunk", "https://s-09.wefunkradio.com:8443/wefunk64.mp3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm72", "Badradio", "https://s2.radio.co/s2b2b68744/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm73", "FIP hiphop", "https://icecast.radiofrance.fr/fiphiphop-midfi.mp3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm74", "HipHop Classics", "https://fluxmusic.api.radiosphere.io/channels/boom-fm-classics/stream.aac", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm75", "Le Rapologue", "https://le-rapologue-radio.radiocult.fm/stream", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm76", "Tequila HipHop", "https://ssl.omegahost.ro/7000/stream", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm77", "DTLR", "https://radio.shoutcheap.com/proxy/shcaesar/stream.mp3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm78", "Itch fm", "https://streaming.radio.co/s264858a04/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm79", "Dilemaradio", "https://dilemaradiolive.radioca.st/stream", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm80", "Base FM", "https://stream.basefm.co.nz/Base320MP3", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm81", "Hot 108 Jamz", "https://live.powerhitz.com/hot108", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm82", "Scratchvision", "https://stream.radio.co/saacde1bcb/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm83", "El Bloke", "https://streaming.radio.co/sd8f1e3a16/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("fm84", "Joint", "https://jointil.com/stream-reggae", "Reggae", "128 kbps"),
-InternetRadioStation("fm85", "Blue radio", "https://a3.asurahosting.com:7960/radio.mp3", "Reggae", "128 kbps"),
-InternetRadioStation("fm86", "TikiPod", "https://cast1.torontocast.com:3330/stream", "Reggae", "128 kbps"),
-InternetRadioStation("fm87", "Roots Legacy", "https://l.rootslegacy.fr/stream", "Reggae", "128 kbps"),
-InternetRadioStation("fm88", "Pigpen", "https://azuracast.spollin.net:8000/radio.mp3", "Reggae", "128 kbps"),
-InternetRadioStation("fm89", "Partyvibe reggae", "https://www.partyviberadio.com:8060", "Reggae", "128 kbps"),
-InternetRadioStation("fm90", "TBN Reggae", "https://puma.streemlion.com:1085/stream", "Reggae", "128 kbps"),
-InternetRadioStation("fm91", "Jah Music Mansion", "https://usa13.fastcast4u.com/proxy/jahmusicmansion?mp=/1", "Reggae", "128 kbps"),
-InternetRadioStation("fm92", "B4B Funk", "https://eu10.fastcast4u.com:8120/;*.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm93", "B-Funk", "https://streams.fluxfm.de/event01/mp3-128/streams.fluxfm.de", "Funk", "128 kbps"),
-InternetRadioStation("fm94", "Soulconnexion", "https://live.soulconnexionradio.com:1045/stream", "Funk", "128 kbps"),
-InternetRadioStation("fm95", "Funky Corner", "https://ais-sa2.cdnstream1.com/2447_192.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm96", "Soul Legends", "https://streamer.radio.co/s60f3e59d1/listen", "Funk", "128 kbps"),
-InternetRadioStation("fm97", "Funky", "https://funkyradio.streamingmedia.it/play.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm98", "Disco Paradise", "https://broadcast.miami/proxy/thediscoparadise?mp=/stream/;", "Funk", "128 kbps"),
-InternetRadioStation("fm99", "Nula", "https://strm.radionula.com/classics", "Funk", "128 kbps"),
-InternetRadioStation("fm100", "Goodfellas Music Station", "https://onair7.xdevel.com/proxy/xautocloud_q1v0_1005?mp=/;stream/", "Funk", "128 kbps"),
-InternetRadioStation("fm101", "Disco fetish", "https://radio.intergalactic.fm/2", "Funk", "128 kbps"),
-InternetRadioStation("fm102", "Luxfunk Blackmix", "https://luxfunkbroadcast.com/proxy/luxfunkradio/stream", "Funk", "128 kbps"),
-InternetRadioStation("fm103", "Palace", "https://www.radioking.com/play/palace-radio", "Funk", "128 kbps"),
-InternetRadioStation("fm104", "Generation Soul Disco Funk", "https://gestream.fr/g-radio-hd.mp3", "Funk", "128 kbps"),
-InternetRadioStation("fm105", "DanceGroove", "https://server7.streamserver24.com:8080/proxy/marasalf?mp=/stream", "Funk", "128 kbps"),
-InternetRadioStation("fm106", "Nicecream Green", "https://play.nicecream.fm/radio/8010/green.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm107", "Bless your sound", "https://ibizasonica.streaming-pro.com:8015/nomorenoless", "Lounge", "128 kbps"),
-InternetRadioStation("fm108", "Cafe Del Mar", "https://streams.radio.co/se1a320b47/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm109", "Croissant Show", "https://streams.radio.co/sddfa96b70/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm110", "Nicecream Red", "https://play.nicecream.fm/radio/8000/red.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm111", "Jazz de Ville Groove", "https://onair22.xdevel.com/proxy/xautocloud_1kha_423?mp=/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm112", "Qmusica", "https://eu8.fastcast4u.com/proxy/qfmserver2?mp=/1", "Lounge", "128 kbps"),
-InternetRadioStation("fm113", "FluxLounge", "https://streams.fluxfm.de/lounge/mp3-128/streams.fluxfm.de", "Lounge", "128 kbps"),
-InternetRadioStation("fm114", "RadioActive", "https://streamyourdream.org:8050/radioactive", "Lounge", "128 kbps"),
-InternetRadioStation("fm115", "Jockey", "https://ec4.yesstreaming.net:2135/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm116", "MotionFM Lounge", "https://vm.motionfm.com/motionthree_aacp", "Lounge", "128 kbps"),
-InternetRadioStation("fm117", "Vanilla Smooth", "https://smooth.vanillaradio.com:8032/live", "Lounge", "128 kbps"),
-InternetRadioStation("fm118", "Nordic Lodge", "https://radio.streemlion.com:3590/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm119", "Crooner", "https://croonerradio.ice.infomaniak.ch/croonerradio-midfi.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm120", "Lounge Radio", "https://nl1.streamhosting.ch/lounge128.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm121", "Salalondon", "https://solid55.streamupsolutions.com/proxy/cfyzzdah/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm122", "Sonicalm", "https://ibizasonica.streaming-pro.com:8014/sonicalm", "Lounge", "128 kbps"),
-InternetRadioStation("fm123", "Skylab", "https://uksoutha.streaming.broadcast.radio:29690/skylab-radio-limited", "Lounge", "128 kbps"),
-InternetRadioStation("fm124", "DinamoFM Caffe", "https://channels.dinamo.fm/caffe-mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm125", "Skyline radio e soul", "https://rblive.it:8010/radio.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm126", "ABC Lounge", "https://listen.openstream.co/3139/audio", "Lounge", "128 kbps"),
-InternetRadioStation("fm127", "Luxuriamusic", "https://ice10.securenetsystems.net/LUXOMP3", "Lounge", "128 kbps"),
-InternetRadioStation("fm128", "Melodies", "https://i.streams.ovh:2200/ssl/melodies?mp=/stream", "Lounge", "128 kbps"),
-InternetRadioStation("fm129", "Bellostream", "https://stream.radio.co/s0ac307198/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm130", "OneLuvFM", "https://streamer.radio.co/s34d8d1549/listen", "Lounge", "128 kbps"),
-InternetRadioStation("fm131", "Smoothlounge", "https://smoothjazz.cdnstream1.com/2586_128.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm132", "Maretimo Lounge", "https://s35.derstream.net/lounge.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm133", "SwissGroove", "https://relay1.swissgroove.ch/;*.mp3", "Lounge", "128 kbps"),
-InternetRadioStation("fm134", "Systrum Sistum", "https://systrum.net:8443/SSR1", "Ambient", "128 kbps"),
-InternetRadioStation("fm135", "Ambient Radio", "https://uk2.internet-radio.com/proxy/ambientradio?mp=/;", "Ambient", "128 kbps"),
-InternetRadioStation("fm136", "ChillHop by FluxFM", "https://streams.fluxfm.de/Chillhop/mp3-128/streams.fluxfm.de", "Ambient", "128 kbps"),
-InternetRadioStation("fm137", "Yoga Sounds by FluxFM", "https://streams.fluxfm.de/yogasounds/mp3-128/streams.fluxfm.de", "Ambient", "128 kbps"),
-InternetRadioStation("fm138", "Verdure", "https://verdure.net/radio/192", "Ambient", "128 kbps"),
-InternetRadioStation("fm139", "Journeyscapes", "https://eagle.streemlion.com/proxy/journeys/stream", "Ambient", "128 kbps"),
-InternetRadioStation("fm140", "Psyndora", "https://cast.magicstreams.gr/sc/psychill/stream", "Ambient", "128 kbps"),
-InternetRadioStation("fm141", "Meuh", "https://radiomeuh2.ice.infomaniak.ch/radiomeuh2-128.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm142", "Noods", "https://noods-radio.radiocult.fm/stream", "Music", "128 kbps"),
-InternetRadioStation("fm143", "Kiosk", "https://kioskradiobxl.out.airtime.pro/kioskradiobxl_b", "Music", "128 kbps"),
-InternetRadioStation("fm144", "Stegi Radio - The Other Side", "https://movementathens2.out.airtime.pro/movementathens2_a", "Music", "128 kbps"),
-InternetRadioStation("fm145", "Le Son Parisien", "https://stream.lesonparisien.com/hi.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm146", "Athens Voice Electronic", "https://n09.radiojar.com/u52ufw1qayduv", "Music", "128 kbps"),
-InternetRadioStation("fm147", "Open Broadcast", "https://stream.openbroadcast.ch/320.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm148", "The Music Machine", "https://msmn3.co/proxy/x2mp33/stream", "Music", "128 kbps"),
-InternetRadioStation("fm149", "OpenLab", "https://ice04.fluidstream.net/openlab.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm150", "Monkey Bros", "https://n0b.radiojar.com/5tyqrwpgpfeuv", "Music", "128 kbps"),
-InternetRadioStation("fm151", "Raheem", "https://radioraheem.out.airtime.pro/radioraheem_a", "Music", "128 kbps"),
-InternetRadioStation("fm152", "Space dogs", "https://www.radioking.com/play/space-dogs-radio", "Music", "128 kbps"),
-InternetRadioStation("fm153", "Vintage Obscura", "https://radio.vintageobscura.net/stream", "Music", "128 kbps"),
-InternetRadioStation("fm154", "DinamoFM Smog", "https://channels.dinamo.fm/smog-mp3", "Music", "128 kbps"),
-InternetRadioStation("fm155", "Stegi Radio", "https://movementathens.out.airtime.pro/movementathens_a", "Music", "128 kbps"),
-InternetRadioStation("fm156", "Music Box", "https://uk5.internet-radio.com/proxy/musicboxradio?mp=/stream", "Music", "128 kbps"),
-InternetRadioStation("fm157", "Global Funk Radio", "https://securestreams2.autopo.st:1148/;*.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm158", "Movie Radio", "https://stream.klassikradio.de/movie-national/aac-128/konsoleplayer", "Music", "128 kbps"),
-InternetRadioStation("fm159", "I Have A Dream", "https://live.radio-ihaveadream.com/stream", "Music", "128 kbps"),
-InternetRadioStation("fm160", "Radio 3S", "https://radio.solarsoundsystem.org/listen/radio3s/radio.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm161", "Worldwide FM", "https://worldwidefm.out.airtime.pro/worldwidefm_b", "Music", "128 kbps"),
-InternetRadioStation("fm162", "Kiss FM", "https://s9.nexuscast.com:9565/kissfm.aac", "Music", "128 kbps"),
-InternetRadioStation("fm163", "Futuradios", "https://futuradiohits.ice.infomaniak.ch/frhits-128.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm164", "Swisspop", "https://stream.srg-ssr.ch/m/rsp/mp3_128", "Music", "128 kbps"),
-InternetRadioStation("fm165", "Maxxima", "https://maxxima.mine.nu/maxxima.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm166", "Frequence3", "https://frequence3.net-radio.fr/frequence3-128.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm167", "PurelectroRadio", "https://str0.creacast.com/purelectro", "Music", "128 kbps"),
-InternetRadioStation("fm168", "EDM Sessions", "https://s2.radio.co/s30844a0f4/listen", "Music", "128 kbps"),
-InternetRadioStation("fm169", "GBH Music", "https://rocket.streamradio.fr:8062/stream", "Music", "128 kbps"),
-InternetRadioStation("fm170", "Tsugi", "https://www.radioking.com/play/tsugi-radio", "Music", "128 kbps"),
-InternetRadioStation("fm171", "8radio", "https://securestreams.autopo.st:1035/;*.mp3", "Music", "128 kbps"),
-InternetRadioStation("fm172", "Le Village Pop", "https://www.radioking.com/play/levillagepop", "Music", "128 kbps"),
-InternetRadioStation("fm173", "Beatles radio", "https://strw3.openstream.co/981", "Music", "128 kbps"),
-InternetRadioStation("fm174", "Dream Machine", "https://radio.intergalactic.fm/3", "Music", "128 kbps"),
-InternetRadioStation("fm175", "NTS", "https://stream-relay-geo.ntslive.net/stream", "Music", "128 kbps"),
-InternetRadioStation("fm176", "17bis", "https://listen.radioking.com/radio/444/stream/1754", "Music", "128 kbps"),
-InternetRadioStation("fm177", "EFR12", "https://listen.radioking.com/radio/151768/stream/192217", "Music", "128 kbps"),
-InternetRadioStation("fm178", "Dublab", "https://dublab.out.airtime.pro/dublab_a", "Music", "128 kbps"),
-InternetRadioStation("fm179", "GlitterBeam", "https://ukwesta.streaming.broadcast.radio/glitterbeam", "Music", "128 kbps"),
-InternetRadioStation("fm180", "Explorers Emporium", "https://www.radioking.com/play/explorers-emporium-radio", "Music", "128 kbps"),
-InternetRadioStation("fm181", "Electro Swing Revolution", "https://streamer.radio.co/s2c3cc784b/listen", "Music", "128 kbps"),
-InternetRadioStation("fm182", "Planet Pootwaddle", "https://ppw.streamguys1.com/sgplayer-mp3", "Music", "128 kbps"),
-InternetRadioStation("fm183", "80s Forever", "https://premium.shoutcastsolutions.com/radio/8050/256.mp3", "Retro", "128 kbps"),
-InternetRadioStation("fm184", "Lolli Oldies", "https://stream.lolliradio.net/lolli_oldies.mp3", "Retro", "128 kbps"),
-InternetRadioStation("fm185", "Fantasy radio", "https://italo.italo.nu/live", "Retro", "128 kbps"),
-InternetRadioStation("fm186", "Groovy Reflections", "https://usa12.fastcast4u.com/proxy/reflecti?mp=/1", "Retro", "128 kbps"),
-InternetRadioStation("fm187", "Retrosounds", "https://listen.radioking.com/radio/34544/stream/70739", "Retro", "128 kbps"),
-InternetRadioStation("fm188", "Stad den Haag", "https://stream.radiostaddenhaag.com/stream/1/;", "Retro", "128 kbps"),
-InternetRadioStation("fm189", "TopSide", "https://listen.radioking.com/radio/33560/stream/69731", "Retro", "128 kbps"),
-InternetRadioStation("fm190", "Affinity", "https://stream.radio.co/s038d07e75/listen", "Retro", "128 kbps"),
-InternetRadioStation("fm191", "TrendFM Den Haag", "https://trendfm.live-streams.nl:18030/live", "Retro", "128 kbps"),
-InternetRadioStation("fm192", "neoFM", "https://streams.fluxfm.de/neofm/mp3-128/streams.fluxfm.de", "Classical", "128 kbps"),
-InternetRadioStation("fm193", "Cinemix", "https://kathy.torontocast.com:1825/stream", "Classical", "128 kbps"),
-InternetRadioStation("fm194", "Minnesota Public Radio", "https://cms.stream.publicradio.org/cms.aac", "Classical", "128 kbps"),
-InternetRadioStation("fm195", "Stephansdom", "https://radioklassikstephansdom.ice.infomaniak.ch/radioklassikstephansdom.mp3", "Classical", "128 kbps"),
-InternetRadioStation("fm196", "Bru Zane", "https://uk2.streamingpulse.com/ssl/bzr", "Classical", "128 kbps"),
-InternetRadioStation("fm197", "Ancientfm", "https://mediaserv73.live-streams.nl:18058/stream", "Classical", "128 kbps"),
-InternetRadioStation("fm198", "Swiss Classic", "https://stream.srg-ssr.ch/m/rsc_de/mp3_128", "Classical", "128 kbps"),
-InternetRadioStation("fm199", "Venice classic", "https://uk2.streamingpulse.com/ssl/vcr1", "Classical", "128 kbps"),
-InternetRadioStation("fm200", "Radio Classique", "https://radioclassique.ice.infomaniak.ch/radioclassique-high.mp3", "Classical", "128 kbps"),
-InternetRadioStation("fm201", "Klassik Radio", "https://stream.klassikradio.de/klr-krdnational/aac-128/konsoleplayer", "Classical", "128 kbps"),
-InternetRadioStation("fm202", "RCO Live", "https://ice.cr6.streamzilla.xlcdn.com:8000/sz=RCOLiveWebradio=mp3-192", "Classical", "128 kbps"),
-InternetRadioStation("fm203", "Accent 4", "https://str0.creacast.com/accent4", "Classical", "128 kbps"),
-InternetRadioStation("fm204", "Schwarzenstein", "https://streams.fluxfm.de/jazzschwarz/mp3-128/streams.fluxfm.de", "Jazz", "128 kbps"),
-InternetRadioStation("fm205", "Concertzender", "https://streams.greenhost.nl:8006/live", "Jazz", "128 kbps"),
-InternetRadioStation("fm206", "Giants of Jazz", "https://streaming.radio.co/s297e618a7/listen", "Jazz", "128 kbps"),
-InternetRadioStation("fm207", "Megaton Cafe", "https://us2.internet-radio.com/proxy/megatoncafe?mp=/stream;", "Jazz", "128 kbps"),
-InternetRadioStation("fm208", "Swing Street", "https://s1.voscast.com:10413/stream", "Jazz", "128 kbps"),
-InternetRadioStation("fm209", "Couleurs Jazz", "https://listen.radioking.com/radio/127546/stream/167344", "Jazz", "128 kbps"),
-InternetRadioStation("fm210", "Smoothjazz", "https://smoothjazz.cdnstream1.com/2585_128.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm211", "Radio Swiss Jazz", "https://stream.srg-ssr.ch/m/rsj/mp3_128", "Jazz", "128 kbps"),
-InternetRadioStation("fm212", "Smooth Jazz Florida", "https://usa6.fastcast4u.com/proxy/wsjfhd?mp=/1", "Jazz", "128 kbps"),
-InternetRadioStation("fm213", "WRTI", "https://wrti-live.streamguys1.com/jazz-mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm214", "TSF 89.9", "https://tsfjazz.ice.infomaniak.ch/tsfjazz-high.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm215", "Jazzradio", "https://jazzradio.ice.infomaniak.ch/jazzradio-high.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("fm216", "QCIndie", "https://stream.qcindie.com/live", "Rock", "128 kbps"),
-InternetRadioStation("fm217", "Loaded", "https://s2.radio.co/sad782b7f8/listen", "Rock", "128 kbps"),
-InternetRadioStation("fm218", "XRaydio", "https://xraydio.ddns.net/live", "Rock", "128 kbps"),
-InternetRadioStation("fm219", "Cathedral 13", "https://usa15.fastcast4u.com/proxy/tstamate?mp=/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm220", "ZZROCK", "https://axtran.radioca.st/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm221", "Wigwam", "https://streaming.radio.co/s2b39cfe62/listen", "Rock", "128 kbps"),
-InternetRadioStation("fm222", "Spoon", "https://spoonradio.ice.infomaniak.ch/spoonradio-light.aac", "Rock", "128 kbps"),
-InternetRadioStation("fm223", "SA Rock", "https://sarock.radioca.st/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm224", "DKFM Shoegaze", "https://kathy.torontocast.com:2005/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm225", "Cosmic Fuzz", "https://15113.live.streamtheworld.com/SAM04AAC335_SC", "Rock", "128 kbps"),
-InternetRadioStation("fm226", "Caroline", "https://stream.radiocaroline.net/;*.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm227", "Surf Rock", "https://radio.dripfeed.net/radio/8020/radio.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm228", "Krautrock world", "https://krautrockworld.stream.laut.fm/krautrockworld", "Rock", "128 kbps"),
-InternetRadioStation("fm229", "Morow", "https://stream.fr.morow.com/morow_med.aacp", "Rock", "128 kbps"),
-InternetRadioStation("fm230", "Bluesradio", "https://i4.streams.ovh/sc/bluesrad/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm231", "Delicious Agony", "https://deliciousagony.streamguys1.com", "Rock", "128 kbps"),
-InternetRadioStation("fm232", "Rock Melodic", "https://i4.streams.ovh:2200/ssl/rockmelo?mp=/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm233", "OUI FM", "https://ouifm.ice.infomaniak.ch/ouifm-high.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm234", "Idobi Howl", "https://idobihowl.idobi.com/;stream.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm235", "12XU", "https://12xu.stream.laut.fm/12xu", "Rock", "128 kbps"),
-InternetRadioStation("fm236", "Free Phoenix", "https://s6.reliastream.com/proxy/radiofre?mp=/stream", "Rock", "128 kbps"),
-InternetRadioStation("fm237", "La Grosse Radio Rock", "https://hd.lagrosseradio.info/lagrosseradio-rock-192.mp3", "Rock", "128 kbps"),
-InternetRadioStation("fm238", "Paul in Rio", "https://usa10.fastcast4u.com/paulinrio/;*.mp3", "World", "128 kbps"),
-InternetRadioStation("fm239", "Exotica", "https://solid67.streamupsolutions.com/proxy/repadnom?mp=/;type=mp3", "World", "128 kbps"),
-InternetRadioStation("fm240", "ILoveRio", "https://eu10.fastcast4u.com/iloverio", "World", "128 kbps"),
-InternetRadioStation("fm241", "Salseo", "https://stream.radio.co/s579ede15c/listen", "World", "128 kbps"),
-InternetRadioStation("gh1", "313.FM Detroit", "http://icecast.ofdoom.com:8000/burst.mp3", "Electronic", "128 kbps"),
-InternetRadioStation("gh2", "AH.FM - EDM Radio", "http://nl.ah.fm:8000/live", "Electronic", "128 kbps"),
-InternetRadioStation("gh3", "9128.live Ambient", "http://streams.radio.co:80/s0aa1e6f4a/listen", "Ambient", "128 kbps"),
-InternetRadioStation("gh4", "A.M. Ambient", "http://radio.stereoscenic.com/ama-h", "Ambient", "128 kbps"),
-InternetRadioStation("gh5", "Ambient Sleeping Pill", "http://radio.stereoscenic.com/asp-h", "Ambient", "128 kbps"),
-InternetRadioStation("gh6", "Bluemars - Cryosleep", "http://50.116.12.253:8000/cryosleep", "Ambient", "128 kbps"),
-InternetRadioStation("gh7", "Bluemars - Voices", "http://50.116.12.253:8000/voicesfromwithin", "Ambient", "128 kbps"),
-InternetRadioStation("gh8", "Bluemars", "http://50.116.12.253:8000/bluemars", "Ambient", "128 kbps"),
-InternetRadioStation("gh9", "Dark Ambient Radio", "http://s3.viastreaming.net:8835/", "Ambient", "128 kbps"),
-InternetRadioStation("gh10", "BBC Radio 1", "http://as-hls-ww-live.akamaized.net/pool_01505109/live/ww/bbc_radio_one/bbc_radio_one.isml/bbc_radio_one-audio%3d96000.norewind.m3u8", "Music", "96 kbps"),
-InternetRadioStation("gh11", "BBC Radio 1Xtra", "http://as-hls-ww-live.akamaized.net/pool_92079267/live/ww/bbc_1xtra/bbc_1xtra.isml/bbc_1xtra-audio%3d96000.norewind.m3u8", "Music", "96 kbps"),
-InternetRadioStation("gh12", "BBC Radio 1Dance", "http://as-hls-ww-live.akamaized.net/pool_62063831/live/ww/bbc_radio_one_dance/bbc_radio_one_dance.isml/bbc_radio_one_dance-audio%3d96000.norewind.m3u8", "Music", "96 kbps"),
-InternetRadioStation("gh13", "BBC Radio 2", "http://as-hls-ww-live.akamaized.net/pool_74208725/live/ww/bbc_radio_two/bbc_radio_two.isml/bbc_radio_two-audio%3d96000.norewind.m3u8", "Music", "96 kbps"),
-InternetRadioStation("gh14", "BBC Radio 6 Music", "http://as-hls-ww-live.akamaized.net/pool_81827798/live/ww/bbc_6music/bbc_6music.isml/bbc_6music-audio%3d96000.norewind.m3u8", "Rock", "96 kbps"),
-InternetRadioStation("gh15", "BBC World Service", "http://a.files.bbci.co.uk/media/live/manifesto/audio/simulcast/hls/nonuk/sbr_low/ak/bbc_world_service.m3u8", "News", "128 kbps"),
-InternetRadioStation("gh16", "FIP", "http://direct.fipradio.fr/live/fip-midfi.mp3", "Music", "128 kbps"),
-InternetRadioStation("gh17", "FIP Jazz", "http://direct.fipradio.fr/live/fip-webradio2.mp3", "Jazz", "128 kbps"),
-InternetRadioStation("gh18", "FIP Groove", "http://direct.fipradio.fr/live/fip-webradio3.mp3", "Music", "128 kbps"),
-InternetRadioStation("gh19", "France Musique", "http://direct.francemusique.fr/live/francemusique-midfi.mp3", "Classical", "128 kbps"),
-InternetRadioStation("gh20", "SomaFM - Groove Salad", "http://ice2.somafm.com/groovesalad-128-aac", "Ambient", "128 kbps"),
-InternetRadioStation("gh21", "SomaFM - Drone Zone", "http://ice2.somafm.com/dronezone-128-aac", "Ambient", "128 kbps"),
-InternetRadioStation("gh22", "SomaFM - Deep Space One", "http://ice2.somafm.com/deepspaceone-128-aac", "Ambient", "128 kbps"),
-InternetRadioStation("gh23", "SomaFM - Space Station", "http://ice2.somafm.com/spacestation-128-aac", "Ambient", "128 kbps"),
-InternetRadioStation("gh24", "SomaFM - Lush", "http://ice2.somafm.com/lush-128-aac", "Electronic", "128 kbps"),
-InternetRadioStation("gh25", "SomaFM - Illinois Street Lounge", "http://ice2.somafm.com/illstreet-128-aac", "Music", "128 kbps"),
-InternetRadioStation("gh26", "KEXP Seattle", "http://live-mp3-128.kexp.org", "Rock", "128 kbps"),
-InternetRadioStation("gh27", "Dublab Los Angeles", "https://dublab.out.airtime.pro/dublab_a", "Music", "128 kbps"),
-InternetRadioStation("gh28", "NTS Radio 1", "http://stream-relay-geo.ntslive.net/stream", "Music", "128 kbps"),
-InternetRadioStation("gh29", "NTS Radio 2", "http://stream-relay-geo.ntslive.net/stream2", "Music", "128 kbps"),
-InternetRadioStation("gh30", "Newtown Radio", "https://streaming.radio.co/s0d090ee43/listen", "Music", "128 kbps"),
-InternetRadioStation("gh31", "Badradio Phonk", "https://s2.radio.co/s2b2b68744/listen", "Hip Hop", "128 kbps"),
-InternetRadioStation("gh32", "Cashmere Radio", "https://cashmereradio.out.airtime.pro/cashmereradio_b", "Music", "128 kbps"),
-InternetRadioStation("gh33", "Kiosk Radio Brussels", "https://kioskradiobxl.out.airtime.pro/kioskradiobxl_b", "Music", "128 kbps"),
-InternetRadioStation("gh34", "NASA Third Rock Radio", "http://rfcm.streamguys1.com/thirdrock-icy", "Music", "128 kbps"),
-InternetRadioStation("gh35", "Nightwave Plaza", "http://radio.plaza.one/ogg", "Music", "128 kbps"),
-InternetRadioStation("gh36", "Radio Paradise", "http://stream.radioparadise.com/global-192", "Music", "192 kbps"),
-InternetRadioStation("gh37", "Radio Caroline", "http://78.129.202.10:8030/", "Music", "128 kbps"),
-InternetRadioStation("gh38", "The Lot Radio NYC", "https://thelot.out.airtime.pro/thelot_a", "Music", "128 kbps"),
-InternetRadioStation("gh39", "Worldwide FM", "https://worldwidefm.out.airtime.pro/worldwidefm_a", "Music", "128 kbps")
-        )
+        _sampleStations.value = buildSampleStations()
     }
 
     fun togglePlayback() {
